### 2018-06-27 #232 @DomBennett
restez -- submission
_The restez package downloads all or sections of GenBank and creates a local SQLite copy of the database for querying. The package comes with a series of useful functions for querying the database and is designed to work with [rentrez](https://github.com/ropensci/rentrez)._

```
Package: restez
Title: Create and Query a Local Copy of GenBank in R
Version: 0.0.0
Authors@R: person("Dom", "Bennett", role = c("aut", "cre"), email = "dominic.john.bennett@gmail.com")
Maintainer: D.J. Bennett <dominic.john.bennett@gmail.com>
Description: Download large sections of GenBank and generate a local SQL-based
    database. A user can then query this database using restez functions or
    through rentrez wrappers.
    rentrez,
    RSQLite,
    DBI,
    R.utils,
    downloader,
    RCurl,
    cli,
    crayon
```

_data retrieval, for users that wish to retrieve lots of sequence information and find NCBI's Entrez too slow._

_Researchers wishing to perform any form of analysis with DNA sequence data. For my own purposes, I will use the package to retrieve large amounts of sequence data for phylogenetic analysis._

_Hajk-Georg Drost's [Biomartr](https://github.com/ropensci/biomartr) is similar and is in fact the inspiration for restez. It only allows users to download genome specific data, however, not GenBank sequences. From the ecologcial sciences perspective, genome data is just not nearly taxonomically representative enough yet for any questions concerning biodiversity._

_NA_

_I imagine the rentrez and biomartr developers to be good reviewers: [dwinter](https://github.com/dwinter) and [HajkD](https://github.com/HajkD)_
