### 2018-07-02 536b4ee
Minor ROpenSci editor check changes
