+++ tests/testthat/test-setup.R
@@ -40,7 +40,9 @@ test_that('db_download() works', {
     `restez:::check_connection` = function() TRUE,
     `restez:::identify_latest_genbank_release_notes` = function() 1,
     `RCurl::getURL` = function(url) '',
-    `restez:::identify_downloadable_files` = function(release_notes) mck_dwnldbl,
+    `restez:::identify_downloadable_files` = function(release_notes) {
+      mck_dwnldbl
+      },
     `restez:::restez_rl` = function(prompt) '1',
     `restez:::file_download` = function(...) TRUE,
     restez:::db_download()
