### 2018-07-03 #232 @naupaka
@sckott I'd be happy to review this if you like. Right in my wheelhouse and looks like something I would make use of for my own research/teaching.
