### 2018-07-03 #232 @sckott
@naupaka thanks for offering, that would be great, thanks, is 3 weeks okay?
