### 2018-07-03 #232 @naupaka
Yes that should work. 👍🏼
