### 2018-07-09 47e26a1
Update paper.md
