### 2018-07-09 5fe0db2
Update paper.md
