### 2018-08-01 #232 @noamross
👋Hi all! Per Evan's comment above, I had suggested using `MonetDBlite` as a mostly drop-in replacement for `RSQLite`.   It's another DBI-compliant back-end stored on-disk. An example of a similar package using it is https://github.com/cboettig/taxald
