### 2018-08-02 #232 @DomBennett
Thanks for such a detailed review @eveskew! I really appreciate the hours you have put into trying and testing the package. This whole ROpenSci malarky is pretty nifty.

@noamross, I will investigate using `MonetDBlite`. It looks like the `taxald` package performs a similar role to restez but for taxonomic information. I have thought about adapting `restez` to also work with NCBI taxonomic information, perhaps this could be in conjunction with `taxald` in the future? There is also the package `taxizedb`, is this significantly different from `taxald`? On the face of it, they both seem to perform the same thing.
