### 2018-08-02 #232 @cboettig
Correct. @DomBennett `taxizedb` downloads the various MySQL / Postgres / SQLite / and tsv dumps from ITIS, NCBI, WikiData, GBIF, COL, and EOL and sticks them into a local version of said database for you.

Each of those databases uses an entirely different format / schema for organizing the taxonomic information, so a query that works with, say, COL to return classification won't work with ITIS, etc.

`taxald` is something of a wrapper around `taxizedb`: it uses taxizedb to download each of these sources.  It then applies a bunch of dplyr wrangling to convert these into a few standard [schemas](https://github.com/cboettig/taxald/blob/master/schema.md).  It then exports the data out of these databases into compressed flat files (using `arkdb`), and uploads them to GitHub (as assets, using `piggyback`... eventually this will be Zenodo once the schemas are a bit more finalized).  That all happens in `data-raw` scripts, since the user doesn't need to repeat that manually each time.

On the user-facing side, `taxald` downloads these compressed flat files from GitHub and loads them into a local MonetDBLite database.  `taxald` user-facing functions are then just relatively simple dplyr queries to the local database.  This database will persist between sessions, and should also makes the queries possible because some of the taxonomic source data are 1 - 6 GB uncompressed, which is probably asking for more RAM than would be polite.
