### 2018-08-02 #232 @cboettig
@DomBennett Very cool package by the way, don't want to interfere with the reviews but looks very well implemented!  In general I'm excited to see more packages that interact with local databases instead of needing millions of API calls.

Completely minor note here, but the way you read in the compressed flat files into the database is pretty slick, though I'm pretty sure you could skip the `utils::gunzip()` since `readLines` can parse a compressed connection directly (gzip, bz2, or xz).  Saves a bit of logic and and a bit of disk space, though maybe not worth changing at this point.  Been thinking about similar problem: `arkdb` uses this logic to import large txt dumps into a database by chunks (since the whole file may exceed memory).
