### 2018-08-17 #232 @naupaka
I’m sorry this has taken so long. I’m just about done putting the package through its paces and looking through the code. I’ve set aside time tomorrow afternoon to finalize the write up.
