### 2018-08-20 #232 @DomBennett
Thanks very much for all the time you put in to this review @naupaka! I'm sorry, it seems my R packages never work for you. It looks like the first error is to do with my expectations for the GenBank record: I suspect the parsing functions are extracting empty strings or unexpected characters/numbers. The `string or blob too big` error is a slightly more worrying. It suggests an upper limit on the length of the sequence. These sorts of errors will have to be looked into afresh when I switch to `MonetDBlite` as suggested above.

@sckott I'll get to addressing the reviewers comments next week.
