### 2018-08-31 6855dc6
Began conversion to monetdb
