+++ man/db_delete.Rd
@@ -26,8 +26,9 @@ db_delete(everything = TRUE)
 (restez_path_get())
 }
 \seealso{
-Other setup: \code{\link{db_create}},
-  \code{\link{db_download}}, \code{\link{demo_db_create}},
+Other setup: \code{\link{connect}},
+  \code{\link{db_create}}, \code{\link{db_download}},
+  \code{\link{demo_db_create}}, \code{\link{disconnect}},
   \code{\link{restez_path_get}},
   \code{\link{restez_path_set}},
   \code{\link{restez_path_unset}},
