+++ man/db_download.Rd
@@ -32,8 +32,9 @@ db_download()
 }
 }
 \seealso{
-Other setup: \code{\link{db_create}},
-  \code{\link{db_delete}}, \code{\link{demo_db_create}},
+Other setup: \code{\link{connect}},
+  \code{\link{db_create}}, \code{\link{db_delete}},
+  \code{\link{demo_db_create}}, \code{\link{disconnect}},
   \code{\link{restez_path_get}},
   \code{\link{restez_path_set}},
   \code{\link{restez_path_unset}},
