+++ man/restez_path_set.Rd
@@ -25,10 +25,10 @@ restez_path_set(filepath = 'path/to/where/you/want/files/to/download')
 }
 }
 \seealso{
-Other setup: \code{\link{db_create}},
-  \code{\link{db_delete}}, \code{\link{db_download}},
-  \code{\link{demo_db_create}},
-  \code{\link{restez_path_get}},
+Other setup: \code{\link{connect}},
+  \code{\link{db_create}}, \code{\link{db_delete}},
+  \code{\link{db_download}}, \code{\link{demo_db_create}},
+  \code{\link{disconnect}}, \code{\link{restez_path_get}},
   \code{\link{restez_path_unset}},
   \code{\link{restez_ready}}, \code{\link{restez_status}}
 }
