+++ R/gb-get-tools.R
@@ -12,9 +12,8 @@ gb_sql_query <- function(nm, id) {
   qry_id <- paste0('(', paste0(paste0("'", id, "'"), collapse = ','), ')')
   qry <- paste0("SELECT accession,", nm,
                 " FROM nucleotide WHERE accession IN ", qry_id)
-  connection <- DBI::dbConnect(drv = MonetDBLite::MonetDBLite(),
-                               dbname = sql_path_get())
-  on.exit(DBI::dbDisconnect(conn = connection))
+  connection <- connect()
+  on.exit(disconnect(connection))
   qry_res <- DBI::dbSendQuery(conn = connection, statement = qry)
   on.exit(expr = {
     DBI::dbClearResult(res = qry_res)
