### 2018-09-03 e14573e
Fixed. Created new connect and disconnect functions
