+++ examples/restez_ready.R
@@ -1,5 +1,7 @@
 library(restez)
 fp <- tempdir()
 restez_path_set(filepath = fp)
+restez_connect()
 demo_db_create()
 (restez_ready())
+restez_disconnect()
