+++ examples/is_in_db.R
@@ -1,6 +1,7 @@
 library(restez)
 # set the restez path to a temporary dir
 restez_path_set(filepath = tempdir())
+restez_connect()
 # create demo database
 demo_db_create(n = 100)
 # in the demo, IDs are 'demo_1', 'demo_2' ...
@@ -9,4 +10,5 @@ ids <- c('thisisnotanid', 'demo_1', 'demo_2')
 
 
 # delete demo after example
+restez_disconnect()
 db_delete()
