+++ examples/restez_status.R
@@ -1,7 +1,9 @@
 library(restez)
 fp <- tempdir()
 restez_path_set(filepath = fp)
+restez_connect()
 demo_db_create()
 restez_status()
+restez_disconnect()
 db_delete()
 restez_status()
