+++ examples/list_db_ids.R
@@ -1,5 +1,6 @@
 library(restez)
 restez_path_set(filepath = tempdir())
+restez_connect()
 demo_db_create(n = 10)
 # Warning: not recommended for real databases
 #  with potentially millions of IDs
@@ -17,4 +18,5 @@ cat(fasta_file)
 
 
 # delete after example
+restez_disconnect()
 db_delete()
