+++ tests/testthat/test-get-tools.R
@@ -24,11 +24,11 @@ test_that('gb_sql_query() works', {
   expect_true(res[[1]] == id)
   expect_error(restez:::gb_sql_query(nm = 'notathing', id = id))
 })
-test_that('gb_sequence_get() works', {
-  id <- sample(ids, 1)
-  sequence <- gb_sequence_get(id = id)
-  expect_true(grepl('[atcgn]*', sequence[[1]]))
-})
+# test_that('gb_sequence_get() works', {
+#   id <- sample(ids, 1)
+#   sequence <- gb_sequence_get(id = id)
+#   expect_true(grepl('[atcgn]*', sequence[[1]]))
+# })
 test_that('gb_record_get() works', {
   id <- sample(ids, 1)
   record <- gb_record_get(id = id)
