### 2018-09-03 03f1f0f
fixed examples. created tests for connect and disconnect
