+++ examples/restez_disconnect.R
@@ -3,6 +3,8 @@ fp <- tempdir()
 restez_path_set(filepath = fp)
 restez_connect()
 demo_db_create()
-(restez_ready())
+restez_status()
+# always remember to disconnect from a database when you've finished
 restez_disconnect()
 db_delete()
+restez_status()
