+++ tests/testthat/test-filepath-tools.R
@@ -66,4 +66,31 @@ test_that('restez_ready() works', {
   demo_db_create()
   expect_true(restez_ready())
 })
+test_that('restez_connect() works', {
+  expect_error(restez_connect())
+  restez:::setup()
+  # must disconnect because setup connects automatically
+  restez_disconnect()
+  on.exit(restez:::cleanup())
+  restez_connect()
+  expect_true(is(restez:::connection_get(), 'MonetDBEmbeddedConnection'))
+})
+test_that('restez_disconnect() works', {
+  expect_null(restez_disconnect())
+  restez:::setup()
+  # must disconnect because setup connects automatically
+  restez_disconnect()
+  on.exit(restez:::cleanup())
+  restez_connect()
+  expect_true(is(restez:::connection_get(), 'MonetDBEmbeddedConnection'))
+  restez_disconnect()
+  expect_error(restez:::connection_get())
+})
+test_that('connection_get() works', {
+  restez:::setup()
+  on.exit(restez:::cleanup())
+  expect_true(is(restez:::connection_get(), 'MonetDBEmbeddedConnection'))
+  restez:::cleanup()
+  expect_error(restez:::connection_get())
+})
 restez:::cleanup()
