+++ man/restez_path_set.Rd
@@ -25,9 +25,9 @@ restez_path_set(filepath = 'path/to/where/you/want/files/to/download')
 }
 }
 \seealso{
-Other setup: \code{\link{connection_get}},
-  \code{\link{db_create}}, \code{\link{db_delete}},
-  \code{\link{db_download}}, \code{\link{demo_db_create}},
+Other setup: \code{\link{db_create}},
+  \code{\link{db_delete}}, \code{\link{db_download}},
+  \code{\link{demo_db_create}},
   \code{\link{restez_connect}},
   \code{\link{restez_disconnect}},
   \code{\link{restez_path_get}},
