+++ man/restez_connect.Rd
@@ -7,13 +7,25 @@
 restez_connect()
 }
 \description{
-Safely creates returns a connection to the local database. If
-database connection cannot be made, an error is returned.
+Returns a connection to the local database. If database
+connection cannot be made, an error is returned.
+}
+\examples{
+library(restez)
+fp <- tempdir()
+restez_path_set(filepath = fp)
+# always run connect before creating or querying a database
+restez_connect()
+demo_db_create()
+restez_status()
+restez_disconnect()
+db_delete()
+restez_status()
 }
 \seealso{
-Other setup: \code{\link{connection_get}},
-  \code{\link{db_create}}, \code{\link{db_delete}},
-  \code{\link{db_download}}, \code{\link{demo_db_create}},
+Other setup: \code{\link{db_create}},
+  \code{\link{db_delete}}, \code{\link{db_download}},
+  \code{\link{demo_db_create}},
   \code{\link{restez_disconnect}},
   \code{\link{restez_path_get}},
   \code{\link{restez_path_set}},
