+++ man/dwnld_path_get.Rd
@@ -15,7 +15,8 @@ are stored.
 }
 \seealso{
 Other private: \code{\link{check_connection}},
-  \code{\link{cleanup}}, \code{\link{custom_download}},
+  \code{\link{cleanup}}, \code{\link{connection_get}},
+  \code{\link{custom_download}},
   \code{\link{entrez_fasta_get}},
   \code{\link{entrez_gb_get}},
   \code{\link{extract_accession}},
