+++ man/restez_ready.Rd
@@ -21,11 +21,12 @@ restez_connect()
 demo_db_create()
 (restez_ready())
 restez_disconnect()
+db_delete()
 }
 \seealso{
-Other setup: \code{\link{connection_get}},
-  \code{\link{db_create}}, \code{\link{db_delete}},
-  \code{\link{db_download}}, \code{\link{demo_db_create}},
+Other setup: \code{\link{db_create}},
+  \code{\link{db_delete}}, \code{\link{db_download}},
+  \code{\link{demo_db_create}},
   \code{\link{restez_connect}},
   \code{\link{restez_disconnect}},
   \code{\link{restez_path_get}},
