+++ man/save_download_details.Rd
@@ -11,7 +11,8 @@ Records downloaded file details.
 }
 \seealso{
 Other private: \code{\link{check_connection}},
-  \code{\link{cleanup}}, \code{\link{custom_download}},
+  \code{\link{cleanup}}, \code{\link{connection_get}},
+  \code{\link{custom_download}},
   \code{\link{dwnld_path_get}},
   \code{\link{entrez_fasta_get}},
   \code{\link{entrez_gb_get}},
