### 2018-09-04 e944ecf
Separated seq and info parts of a GB record in db
