+++ tests/testthat/test-entrez-tools.R
@@ -14,7 +14,7 @@ restez:::cleanup()
 restez:::setup()
 df <- restez:::gb_df_generate(records = sample(records, size = nrcrds))
 ids <- as.character(df[['accession']])
-restez:::gb_sql_add(df = df, database = 'nucleotide')
+restez:::gb_sql_add(df = df)
 
 # RUNNING
 context('Testing \'entrez-tools\'')
