### 2018-09-04 3fbb037
Made all explicit returns of NULL invisible
