### 2018-09-04 2474381
Connect disconnect added to doc
