### 2018-09-04 8a63eb9
Session info added to fp
