### 2018-09-04 82ed997
Minor fixes plus version update
