+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: restez
 Type: Package
 Title: Create and Query a Local Copy of GenBank in R
-Version: 0.0.0
+Version: 0.1.0
 Authors@R: person("Dom", "Bennett", role = c("aut", "cre"), email = "dominic.john.bennett@gmail.com")
 Maintainer: D.J. Bennett <dominic.john.bennett@gmail.com>
 Description: Download large sections of GenBank and generate a local SQL-based
@@ -23,7 +23,8 @@ Imports:
     downloader,
     RCurl,
     cli,
-    crayon
+    crayon,
+    devtools
 Suggests:
     testthat,
     knitr,
