+++ R/biomartr-tools.R
@@ -28,27 +28,15 @@ check_connection <- function() {
 custom_download <- function(...) {
   operating_sys <- Sys.info()[1]
   if (operating_sys == "Darwin") {
-    downloader::download(
-      ...,
-      method = "curl",
-      extra = "--connect-timeout 120 --retry 3",
-      cacheOK = FALSE,
-         quiet = TRUE
-    )
+    downloader::download(..., method = "curl", cacheOK = FALSE, quiet = TRUE,
+                         extra = "--connect-timeout 120 --retry 3")
   }
   if (operating_sys == "Linux") {
-    downloader::download(
-      ...,
-      method = "wget",
-      extra = "--timeout 120 --tries 3 --continue",
-      cacheOK = FALSE,
-      quiet = TRUE
-    )
+    downloader::download(..., method = "wget", cacheOK = FALSE, quiet = TRUE,
+                         extra = "--timeout 120 --tries 3 --continue")
   }
   if (operating_sys == "Windows") {
-    downloader::download(...,
-                         method = "internal",
-                         cacheOK = FALSE,
+    downloader::download(..., method = "internal", cacheOK = FALSE,
                          quiet = TRUE)
   }
 }
