### 2018-09-04 98e41f0
Broken. Beginning log tools
