+++ R/download-tools.R
@@ -1,7 +1,8 @@
 #' @name identify_latest_genbank_release_notes
 #' @title Identify the latest GenBank Release Notes
-#' @description Searches through all release notes
-#' to find the latest. Returns the entire release notes.
+#' @description Searches through all release notes to find the latest. Returns
+#' "gb[release number].release.notes".
+#' @param just_release_number Logical. Default FALSE.
 #' @return character
 #' @family private
 identify_latest_genbank_release_notes <- function() {
@@ -103,14 +104,8 @@ file_download <- function(fl, overwrite=FALSE) {
     remove(gzdest)
     FALSE
   })
+  if (success) {
+    download_record_log(fl)
+  }
   success
-}
-
-#' @name save_download_details
-#' @title Save download details
-#' @description Records downloaded file details.
-#' @return NULL
-#' @family private
-save_download_details <- function() {
-
 }
