### 2018-09-05 b32bc99
Added new log functions
