+++ R/log-tools.R
@@ -1,20 +1,67 @@
-download_record_log <- function(fl) {
+# TODO: record user's selection.
+# LOG TOOLS
+#' @name dwnld_rcrd_log
+#' @title Log a downloaded file in the restez path
+#' @description This function is called whenever a file is successfully
+#' downloaded. A row entry is added to the 'download_log.tsv' in the user's
+#' restez path containing the file name, the GB release number and the time of
+#' successfuly download. The log is to help users keep track of when they
+#' downloaded files and to determine if the downloaded files are out of date.
+#' @param fl file name, character
+#' @return NULL
+#' @family private
+dwnld_rcrd_log <- function(fl) {
   fp <- file.path(restez_path_get(), 'download_log.tsv')
-  row_entry <- data.frame('File' = fl, 'GB.release' = release_get(),
+  row_entry <- data.frame('File' = fl, 'GB.release' = gbrelease_get(),
                           'Time' = as.character(Sys.time()))
   write.table(x = row_entry, file = fp, append = file.exists(fp), sep = '\t',
-              col.names = !file.exists(fp), row.names = FALSE)
+              col.names = !file.exists(fp), row.names = FALSE, quote = FALSE)
 }
 
-add_record_log <- function() {
-
+#' @name add_rcrd_log
+#' @title Log sequences added to the SQL database in the restez path
+#' @description This function is called whenever sequences have been
+#' successfully added to the nucleotide SQL database. Row entries are added to
+#' 'add_lot.tsv' in the user's restez path containing the sequence accession,
+#' version and GB release numbers as well as the time of successfuly adding.
+#' The log is to help users keep track of when sequences have been added.
+#' @param df sequence records, data.frame
+#' @return NULL
+#' @family private
+add_rcrd_log <- function(df) {
+  fp <- file.path(restez_path_get(), 'add_log.tsv')
+  rows_entry <- data.frame('Accession' = df[['accession']],
+                           'Version' = df[['version']],
+                           'GB.release' = gbrelease_get(),
+                           'Time' = as.character(Sys.time()))
+  write.table(x = rows_entry, file = fp, append = file.exists(fp), sep = '\t',
+              col.names = !file.exists(fp), row.names = FALSE, quote = FALSE)
 }
 
-gbrelease_log <- function() {
-
+#' @name gbrelease_log
+#' @title Log the GenBank release number in the restez path
+#' @description This function is called whenever db_download is run. It logs the
+#' GB release number in the 'gb_release.txt' in the user's restez path.
+#' The log is to help users keep track of whether their database if out of date.
+#' @param release GenBank release number, character
+#' @return NULL
+#' @family private
+gbrelease_log <- function(release) {
+  # release number can be in the form 'gb.release####'
+  fp <- file.path(restez_path_get(), 'gb_release.txt')
+  write(x = gsub(pattern = '[^0-9]', replacement = '', x = release), file = fp)
 }
 
-release_get <- function() {
+# GET TOOLS
+#' @name gbrelease_log
+#' @title Log the GenBank release number in the restez path
+#' @description This function is called whenever db_download is run. It logs the
+#' GB release number in the 'gb_release.txt' in the user's restez path.
+#' The log is to help users keep track of whether their database if out of date.
+#' @param release GenBank release number, character
+#' @return NULL
+#' @family private
+gbrelease_get <- function() {
   fp <- file.path(restez_path_get(), 'gb_release.txt')
   if (file.exists(fp)) {
     res <- read.table(file = fp, header = FALSE)[[1]]
[…]

[…]
