### 2018-09-05 ddb6552
Passing. Added selection choice log
