+++ man/file_download.Rd
@@ -46,7 +46,7 @@ Other private: \code{\link{add_rcrd_log}},
   \code{\link{gb_df_generate}}, \code{\link{gb_sql_add}},
   \code{\link{gb_sql_query}},
   \code{\link{gbrelease_check}},
-  \code{\link{gbrelease_log}},
+  \code{\link{gbrelease_get}}, \code{\link{gbrelease_log}},
   \code{\link{identify_downloadable_files}},
   \code{\link{identify_latest_genbank_release_notes}},
   \code{\link{last_add_get}}, \code{\link{last_dwnld_get}},
@@ -56,5 +56,6 @@ Other private: \code{\link{add_rcrd_log}},
   \code{\link{mock_org}}, \code{\link{mock_rec}},
   \code{\link{mock_seq}}, \code{\link{restez_path_check}},
   \code{\link{restez_rl}}, \code{\link{setup}},
+  \code{\link{slctn_get}}, \code{\link{slctn_log}},
   \code{\link{sql_path_get}}, \code{\link{testdatadir_get}}
 }
