### 2018-09-05 b4b6f00
New seqlength logging
