+++ R/startup.R
@@ -1,12 +1,15 @@
 .onAttach <- function(...) {
   v <- utils::packageVersion("restez")
+  v_msg <- paste0('restez v', v)
+  v_msg_bar <- paste0(rep(x = '-', nchar(v_msg)), collapse = '')
+  v_msg <- paste0(v_msg_bar, '\n', v_msg, '\n', v_msg_bar, '\n')
   fp <- restez_path_get()
   if (is.null(fp)) {
-    msg <- paste0('restez ', v, '\n',
-                  'Remember to restez_path_set()')
+    msg <- paste0(v_msg, 'Remember to restez_path_set() and, then, ',
+                  'restez_connect()')
   } else {
-    msg <- paste0('restez v', v, '\n',
-                  'path = [', fp, ']\n')
+    msg <- paste0(v_msg, 'Restez path = [', fp,
+                  ']\nRemeber to restez_connect()')
   }
   packageStartupMessage(msg)
 }
