+++ R/log-tools.R
@@ -1,4 +1,20 @@
-# LOG TOOLS
+# LOG TOOLS ----
+#' @name db_sqlngths_log
+#' @title Log the min and max sequence lengths
+#' @description Log the min and maximum sequnece length used in the created db.
+#' @param min_lngth Minimum length
+#' @param max_lngth Maximum length
+#' @return NULL
+#' @family private
+db_sqlngths_log <- function(min_lngth, max_lngth) {
+  fp <- file.path(restez_path_get(), 'seqlengths.tsv')
+  if (is.null(max_lngth)) {
+    max_lngth <- 'Inf'
+  }
+  row_entry <- data.frame('min' = min_lngth, 'max' = max_lngth)
+  utils::write.table(x = row_entry, file = fp, sep = '\t')
+}
+
 #' @name slctn_log
 #' @title Log the GenBank selection made by a user
 #' @description This function is called whenever a user makes a selection with
@@ -70,14 +86,18 @@ gbrelease_log <- function(release) {
   write(x = gsub(pattern = '[^0-9]', replacement = '', x = release), file = fp)
 }
 
-# GET TOOLS
+# GET TOOLS ----
 #' @name slctn_get
 #' @title Retrieve GenBank selections made by user
 #' @description Returns the selections made by the user.
+#' @details If no file found, returns empty character vector.
 #' @return character vector
 #' @family private
 slctn_get <- function() {
   fp <- file.path(restez_path_get(), 'selection_log.tsv')
+  if (!file.exists(fp)) {
+    return('')
+  }
   sort(unique(utils::read.table(file = fp, header = TRUE, sep = '\t',
                                 stringsAsFactors = FALSE)[['Selection']]))
 }
@@ -87,13 +107,14 @@ slctn_get <- function() {
 #' @description Returns the GenBank release number. Returns empty character
 #' if none found.
 #' @return character
+#' @details If no file found, returns empty character vector.
 #' @family private
 gbrelease_get <- function() {
   fp <- file.path(restez_path_get(), 'gb_release.txt')
   if (file.exists(fp)) {
     res <- utils::read.table(file = fp, header = FALSE)[[1]]
   } else {
-    res <- ''
+    res <- '0'
   }
   res
 }
@@ -115,9 +136,13 @@ last_entry_get <- function(fp) {
 #' @description Return the date and time of the last download as determined
 #' using the 'download_log.tsv'.
 #' @return character
+#' @details If no file found, returns empty character vector.
 #' @family private
 last_dwnld_get <- function() {
   fp <- file.path(restez_path_get(), 'download_log.tsv')
+  if (!file.exists(fp)) {
+    return('')
+  }
   last_entry_get(fp = fp)[[3]]
 }
 
@@ -126,9 +151,13 @@ last_dwnld_get <- function() {
 #' @description Return the date and time of the last added sequence as
 #' determined using the 'add_log.tsv'.
 #' @return character
+#' @details If no file found, returns empty character vector.
 #' @family private
 last_add_get <- function() {
   fp <- file.path(restez_path_get(), 'add_log.tsv')
+  if (!file.exists(fp)) {
+    return('')
+  }
   last_entry_get(fp = fp)[[4]]
 }
 
@@ -150,6 +179,24 @@ db_nrows_get <- function() {
   res[[1]]
 }
 
+#' @name db_sqlngths_get
+#' @title Return the minimum and maximum sequence lengths in db
+#' @description Returns the maximum and minimum sequence lengths as set by the
+#' user upon db creation.
+#' @return vector of integers
+#' @details If no file found, returns empty character vector.
+#' @family private
+db_sqlngths_get <- function() {
+  fp <- file.path(restez_path_get(), 'seqlengths.tsv')
+  if (!file.exists(fp)) {
[…]

[…]
