### 2018-09-05 fd2e225
separated session and readme from restez_path_set
