### 2018-09-05 056094c
Fixed non-recognised character error
