### 2018-09-06 c6ed8c6
fasta and sequence cleaning error fixes
