### 2018-09-06 f8b3409
reworked add_log
