+++ man/last_add_get.Rd
@@ -40,9 +40,10 @@ Other private: \code{\link{add_rcrd_log}},
   \code{\link{extract_seqrecpart}},
   \code{\link{extract_sequence}},
   \code{\link{extract_version}},
-  \code{\link{file_download}}, \code{\link{flatfile_read}},
-  \code{\link{gb_df_create}}, \code{\link{gb_df_generate}},
-  \code{\link{gb_sql_add}}, \code{\link{gb_sql_query}},
+  \code{\link{file_download}}, \code{\link{filename_log}},
+  \code{\link{flatfile_read}}, \code{\link{gb_df_create}},
+  \code{\link{gb_df_generate}}, \code{\link{gb_sql_add}},
+  \code{\link{gb_sql_query}},
   \code{\link{gbrelease_check}},
   \code{\link{gbrelease_get}}, \code{\link{gbrelease_log}},
   \code{\link{identify_downloadable_files}},
