### 2018-09-06 69cd204
created subprocess tools for dot spinning and better kill-control
