### 2018-09-06 612eec7
new documentation on new functions
