+++ other/dev2.R
@@ -1,6 +1,0 @@
-
-
-library(restez)
-devtools::load_all('~/Coding/restez')
-restez_path_set('~/Desktop')
-db_download(preselection = '15')
