+++ man/custom_download.Rd
@@ -16,10 +16,15 @@ ftp file downloads are customized for each operating system.
 }
 \seealso{
 Other private: \code{\link{add_rcrd_log}},
+  \code{\link{cat_line}}, \code{\link{char}},
   \code{\link{check_connection}}, \code{\link{cleanup}},
-  \code{\link{connection_get}}, \code{\link{db_nrows_get}},
+  \code{\link{connection_get}},
+  \code{\link{custom_download2}},
+  \code{\link{db_nrows_get}},
   \code{\link{db_sqlngths_get}},
   \code{\link{db_sqlngths_log}}, \code{\link{dir_size}},
+  \code{\link{dot_spinner}},
+  \code{\link{download_cmd_generate}},
   \code{\link{dwnld_path_get}},
   \code{\link{dwnld_rcrd_log}},
   \code{\link{entrez_fasta_get}},
@@ -42,6 +47,7 @@ Other private: \code{\link{add_rcrd_log}},
   \code{\link{gb_sql_query}},
   \code{\link{gbrelease_check}},
   \code{\link{gbrelease_get}}, \code{\link{gbrelease_log}},
+  \code{\link{handle_run}},
   \code{\link{identify_downloadable_files}},
   \code{\link{identify_latest_genbank_release_notes}},
   \code{\link{last_add_get}}, \code{\link{last_dwnld_get}},
@@ -51,9 +57,11 @@ Other private: \code{\link{add_rcrd_log}},
   \code{\link{mock_org}}, \code{\link{mock_rec}},
   \code{\link{mock_seq}}, \code{\link{readme_log}},
   \code{\link{restez_path_check}}, \code{\link{restez_rl}},
+  \code{\link{rhandle_generate}},
   \code{\link{seshinfo_log}}, \code{\link{setup}},
   \code{\link{slctn_get}}, \code{\link{slctn_log}},
-  \code{\link{sql_path_get}}, \code{\link{testdatadir_get}}
+  \code{\link{sql_path_get}}, \code{\link{stat}},
+  \code{\link{testdatadir_get}}
 }
 \author{
 Hajk-Georg Drost

[…]
