### 2018-09-06 ff05d32
new print tools tests
