### 2018-09-06 44ac703
log tool tests
