### 2018-09-07 c35e56d
Made sequence point in record more explicit
