### 2018-09-07 57b5292
updated extract to be more precise. identified records where errors occurred.
