+++ man/mock_org.Rd
@@ -33,7 +33,7 @@ Other private: \code{\link{add_rcrd_log}},
   \code{\link{entrez_fasta_get}},
   \code{\link{entrez_gb_get}},
   \code{\link{extract_accession}},
-  \code{\link{extract_by_keyword}},
+  \code{\link{extract_by_patterns}},
   \code{\link{extract_clean_sequence}},
   \code{\link{extract_definition}},
   \code{\link{extract_features}},
