### 2018-09-07 7ff401e
new tester script
