+++ other/dev.R
@@ -12,3 +12,10 @@ length(ids)
 cat(gb_fasta_get(sample(ids, 2)))
 gb_organism_get(sample(ids, 2))
 cat(gb_record_get(sample(ids, 2)))
+
+# TODO:
+# - update 'records' data with difficult records
+# - create a download_path for db_create
+# - create connected() and with_data() functions
+# - breakup restez_status, output structured list, bring back restez_ready
+#
