### 2018-09-09 4a15440
allowing records without sequence
