### 2018-09-11 768c6c7
return empty list if too few lines in flatfile
