### 2018-09-11 245db6e
updated release notes lookup
