### 2018-09-11 6deb3b8
Test error fixes
