### 2018-09-11 a7fa576
Fixed and added filesizes
