+++ other/dev.R
@@ -4,7 +4,6 @@
 
 # TODO:
 # - update 'records' data with difficult records
-# - create a download_path for db_create
 # - create connected() and with_data() functions
 # - breakup restez_status, output structured list, bring back restez_ready
 #
