+++ man/latest_genbank_release.Rd
@@ -1,17 +1,16 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/download-tools.R
-\name{identify_latest_genbank_release_notes}
-\alias{identify_latest_genbank_release_notes}
-\title{Identify the latest GenBank Release Notes}
+\name{latest_genbank_release}
+\alias{latest_genbank_release}
+\title{Retrieve latest GenBank release number}
 \usage{
-identify_latest_genbank_release_notes()
+latest_genbank_release()
 }
 \value{
 character
 }
 \description{
-Searches through all release notes to find the latest. Returns
-"gb[release number].release.notes".
+Downloads the latest GenBank release number and returns it.
 }
 \seealso{
 Other private: \code{\link{add_rcrd_log}},
@@ -51,6 +50,7 @@ Other private: \code{\link{add_rcrd_log}},
   \code{\link{identify_downloadable_files}},
   \code{\link{last_add_get}}, \code{\link{last_dwnld_get}},
   \code{\link{last_entry_get}},
+  \code{\link{latest_genbank_release_notes}},
   \code{\link{message_missing}}, \code{\link{mock_def}},
   \code{\link{mock_gb_df_generate}},
   \code{\link{mock_org}}, \code{\link{mock_rec}},
