+++ man/gbrelease_log.Rd
@@ -49,9 +49,10 @@ Other private: \code{\link{add_rcrd_log}},
   \code{\link{gbrelease_check}},
   \code{\link{gbrelease_get}}, \code{\link{handle_run}},
   \code{\link{identify_downloadable_files}},
-  \code{\link{identify_latest_genbank_release_notes}},
   \code{\link{last_add_get}}, \code{\link{last_dwnld_get}},
   \code{\link{last_entry_get}},
+  \code{\link{latest_genbank_release_notes}},
+  \code{\link{latest_genbank_release}},
   \code{\link{message_missing}}, \code{\link{mock_def}},
   \code{\link{mock_gb_df_generate}},
   \code{\link{mock_org}}, \code{\link{mock_rec}},
