+++ man/latest_genbank_release_notes.Rd
@@ -1,18 +1,14 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/biomartr-tools.R
-\name{custom_download}
-\alias{custom_download}
-\title{Helper function to perform customized downloads}
+% Please edit documentation in R/download-tools.R
+\name{latest_genbank_release_notes}
+\alias{latest_genbank_release_notes}
+\title{Download the latest GenBank Release Notes}
 \usage{
-custom_download(...)
-}
-\arguments{
-\item{...}{additional arguments that shall be passed to
-\code{\link[downloader]{download}}}
+latest_genbank_release_notes()
 }
 \description{
-To achieve the most stable download experience,
-ftp file downloads are customized for each operating system.
+Downloads the latest GenBank release notes to a user's restez
+download path.
 }
 \seealso{
 Other private: \code{\link{add_rcrd_log}},
@@ -20,6 +16,7 @@ Other private: \code{\link{add_rcrd_log}},
   \code{\link{check_connection}}, \code{\link{cleanup}},
   \code{\link{connection_get}},
   \code{\link{custom_download2}},
+  \code{\link{custom_download}},
   \code{\link{db_nrows_get}},
   \code{\link{db_sqlngths_get}},
   \code{\link{db_sqlngths_log}}, \code{\link{dir_size}},
@@ -51,7 +48,6 @@ Other private: \code{\link{add_rcrd_log}},
   \code{\link{identify_downloadable_files}},
   \code{\link{last_add_get}}, \code{\link{last_dwnld_get}},
   \code{\link{last_entry_get}},
-  \code{\link{latest_genbank_release_notes}},
   \code{\link{latest_genbank_release}},
   \code{\link{message_missing}}, \code{\link{mock_def}},
   \code{\link{mock_gb_df_generate}},
@@ -63,7 +59,4 @@ Other private: \code{\link{add_rcrd_log}},
   \code{\link{slctn_get}}, \code{\link{slctn_log}},
   \code{\link{sql_path_get}}, \code{\link{stat}},
   \code{\link{testdatadir_get}}
-}
-\author{
-Hajk-Georg Drost
 }
