### 2018-09-12 28a11b0
switch to line spinner
