+++ man/check_connection.Rd
@@ -21,7 +21,6 @@ Other private: \code{\link{add_rcrd_log}},
   \code{\link{db_nrows_get}},
   \code{\link{db_sqlngths_get}},
   \code{\link{db_sqlngths_log}}, \code{\link{dir_size}},
-  \code{\link{dot_spinner}},
   \code{\link{download_cmd_generate}},
   \code{\link{dwnld_path_get}},
   \code{\link{dwnld_rcrd_log}},
@@ -51,6 +50,7 @@ Other private: \code{\link{add_rcrd_log}},
   \code{\link{last_entry_get}},
   \code{\link{latest_genbank_release_notes}},
   \code{\link{latest_genbank_release}},
+  \code{\link{line_spinner}},
   \code{\link{message_missing}}, \code{\link{mock_def}},
   \code{\link{mock_gb_df_generate}},
   \code{\link{mock_org}}, \code{\link{mock_rec}},
