+++ man/dir_size.Rd
@@ -24,7 +24,7 @@ Other private: \code{\link{add_rcrd_log}},
   \code{\link{custom_download}},
   \code{\link{db_nrows_get}},
   \code{\link{db_sqlngths_get}},
-  \code{\link{db_sqlngths_log}}, \code{\link{dot_spinner}},
+  \code{\link{db_sqlngths_log}},
   \code{\link{download_cmd_generate}},
   \code{\link{dwnld_path_get}},
   \code{\link{dwnld_rcrd_log}},
@@ -54,6 +54,7 @@ Other private: \code{\link{add_rcrd_log}},
   \code{\link{last_entry_get}},
   \code{\link{latest_genbank_release_notes}},
   \code{\link{latest_genbank_release}},
+  \code{\link{line_spinner}},
   \code{\link{message_missing}}, \code{\link{mock_def}},
   \code{\link{mock_gb_df_generate}},
   \code{\link{mock_org}}, \code{\link{mock_rec}},
