### 2018-09-12 d6365b0
pulled up filepath, created status tools
