+++ other/dev.R
@@ -4,6 +4,9 @@ devtools::load_all('~/Coding/restez')
 restez_path_set('~/Desktop')
 restez_connect()
 restez_status(gb_check = TRUE)
+
+connected()
+
 restez_disconnect()
 db_delete(everything = TRUE)
 
@@ -11,6 +14,4 @@ db_download()
 
 # TODO:
 # - update 'records' data with difficult records
-# - create connected() and with_data() functions
-# - breakup restez_status, output structured list, bring back restez_ready
 # - update file size calulcations
