+++ R/filepath-tools.R
@@ -93,185 +93,3 @@ restez_path_check <- function() {
     stop(msg, .call = FALSE)
   }
 }
-
-#' @name db_delete
-#' @title Delete database
-#' @family setup
-#' @description Delete the local SQL database and/or restez
-#' folder.
-#' @param everything T/F, delete the whole restez folder as well?
-#' @return NULL
-#' @export
-#' @example examples/db_delete.R
-db_delete <- function(everything = TRUE) {
-  if (length(sql_path_get()) > 0 && dir.exists(sql_path_get())) {
-    unlink(sql_path_get(), recursive = TRUE)
-  }
-  if (everything) {
-    if (length(restez_path_get()) > 0 && dir.exists(restez_path_get())) {
-      unlink(restez_path_get(), recursive = TRUE)
-      restez_path_unset()
-    }
-  }
-  invisible(NULL)
-}
-
-#' @name restez_status
-#' @title Check restez status
-#' @family setup
-#' @description Report to console current setup status of restez. If SQL
-#' database available, returns TRUE else FALSE.
-#' @param gb_check Check whether last download was from latest GenBank release?
-#' Default FALSE.
-#' @details If the function returns TRUE, then querying can begin. Always
-#' remember to run \code{\link{restez_connect}} before running this function.
-#' Set gb_check=TRUE to see if your downloads are up-to-date.
-#' @return T/F
-#' @export
-#' @example examples/restez_status.R
-restez_status <- function(gb_check = FALSE) {
-  # TODO: return a structured list
-  no_downloads <- FALSE
-  no_database <- FALSE
-  latest <- TRUE
-  fp <- restez_path_get()
-  cat_line(cli::rule())
-  cat_line('Checking setup status at ', char(fp), ' ...')
-  cat_line(cli::rule())
-  if (is.null(fp)) {
-    cat_line('... restez path not set')
-    message('You need to use restez_path_set()')
-    return(FALSE)
-  }
-  if (!dir.exists(fp)) {
-    cat_line('... restez path ', char(fp), ' does not exist')
-    message('restez_path_set() filepath must be a valid filepath')
-    return(FALSE)
-  }
-  fp <- dwnld_path_get()
-  if (!dir.exists(fp)) {
-    cat_line('... ', char('downloads/'), ' does not exist')
-    message('Use restez_path_set() to recreate the folder')
-    return(FALSE)
-  }
-  dwn_fls <- list.files(path = fp)
-  if (length(dwn_fls) == 0) {
-    cat_line('... no files in ', char('downloads/'))
-    no_downloads <- TRUE
-  } else {
-    dwn_fls <- file.path(fp, dwn_fls)
-    cat_line('... found ', stat(length(dwn_fls)), ' files in ',
-             char('downloads/'))
-    cat_line('... totalling ', stat(dir_size(dwnld_path_get()), 'GB'))
-    cat_line('... of sequences representing:')
-    for (slctn in slctn_get()) {
-      cat_line('... ... ', char(slctn))
-    }
-    cat_line('... last download was made on ', char(last_dwnld_get()))
-    cat_line('... GenBank relase number ', stat(gbrelease_get()))
-    if (gb_check) {
-      latest <- gbrelease_check()
-    }
-  }
-  cat_line(cli::rule())
-  if (!restez_ready()) {
-    cat_line('... ', char('sql_db'), ' does not exist or is not connected')
-    no_database <- TRUE
-  } else {
-    cat_line('... found ', char('sql_db'), ' of ',
-             stat(dir_size(sql_path_get()), 'GB'))
-    cat_line('... and ', stat(db_nrows_get()), ' rows')
-    sqlngths <- db_sqlngths_get()
-    cat_line('... and sequence length limits of ',
-             stat(sqlngths[['min']], 'bp'), ' to ',
[…]

[…]
