### 2018-09-12 4bcd11f
Updated tests, ensured tests function without internet
