### 2018-09-13 7d58f09
fixed test errors. made rentrez and restez output identical
