+++ examples/restez_connect.R
@@ -7,4 +7,5 @@ demo_db_create()
 restez_status()
 restez_disconnect()
 db_delete()
-restez_status()
+# Errors
+# restez_status()
