+++ examples/restez_ready.R
@@ -3,8 +3,7 @@ fp <- tempdir()
 restez_path_set(filepath = fp)
 restez_connect()
 demo_db_create()
-restez_status()
+(restez_ready())
 restez_disconnect()
 db_delete()
-# Errors:
-# restez_status()
+(restez_ready())
