+++ examples/restez_disconnect.R
@@ -7,4 +7,5 @@ restez_status()
 # always remember to disconnect from a database when you've finished
 restez_disconnect()
 db_delete()
-restez_status()
+# Errors
+# restez_status()
