+++ man/predict_datasizes.Rd
@@ -1,17 +1,18 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/log-tools.R
-\name{gbrelease_check}
-\alias{gbrelease_check}
-\title{Check if the last GenBank release number is the latest}
+% Please edit documentation in R/download-tools.R
+\name{predict_datasizes}
+\alias{predict_datasizes}
+\title{Print file size predictions to screen}
 \usage{
-gbrelease_check()
+predict_datasizes(uncompressed_filesize)
 }
-\value{
-logical
+\arguments{
+\item{uncompressed_filesize}{GBs of the stated filesize, numeric}
 }
 \description{
-Returns TRUE if the GenBank release number is the most recent
-GenBank release available.
+Predicts the file sizes of the downloads and the database
+from the GenBank filesize information. Conversion factors are based on
+previous restez downloads.
 }
 \seealso{
 Other private: \code{\link{add_rcrd_log}},
@@ -43,9 +44,10 @@ Other private: \code{\link{add_rcrd_log}},
   \code{\link{file_download}}, \code{\link{filename_log}},
   \code{\link{flatfile_read}}, \code{\link{gb_df_create}},
   \code{\link{gb_df_generate}}, \code{\link{gb_sql_add}},
-  \code{\link{gb_sql_query}}, \code{\link{gbrelease_get}},
-  \code{\link{gbrelease_log}}, \code{\link{handle_run}},
-  \code{\link{has_data}},
+  \code{\link{gb_sql_query}},
+  \code{\link{gbrelease_check}},
+  \code{\link{gbrelease_get}}, \code{\link{gbrelease_log}},
+  \code{\link{handle_run}}, \code{\link{has_data}},
   \code{\link{identify_downloadable_files}},
   \code{\link{last_add_get}}, \code{\link{last_dwnld_get}},
   \code{\link{last_entry_get}},
@@ -55,8 +57,8 @@ Other private: \code{\link{add_rcrd_log}},
   \code{\link{message_missing}}, \code{\link{mock_def}},
   \code{\link{mock_gb_df_generate}},
   \code{\link{mock_org}}, \code{\link{mock_rec}},
-  \code{\link{mock_seq}}, \code{\link{predict_datasizes}},
-  \code{\link{print.status}}, \code{\link{readme_log}},
+  \code{\link{mock_seq}}, \code{\link{print.status}},
+  \code{\link{readme_log}},
   \code{\link{restez_path_check}}, \code{\link{restez_rl}},
   \code{\link{rhandle_generate}},
   \code{\link{seshinfo_log}}, \code{\link{setup}},
