+++ man/message_missing.Rd
@@ -54,8 +54,8 @@ Other private: \code{\link{add_rcrd_log}},
   \code{\link{line_spinner}}, \code{\link{mock_def}},
   \code{\link{mock_gb_df_generate}},
   \code{\link{mock_org}}, \code{\link{mock_rec}},
-  \code{\link{mock_seq}}, \code{\link{print.status}},
-  \code{\link{readme_log}},
+  \code{\link{mock_seq}}, \code{\link{predict_datasizes}},
+  \code{\link{print.status}}, \code{\link{readme_log}},
   \code{\link{restez_path_check}}, \code{\link{restez_rl}},
   \code{\link{rhandle_generate}},
   \code{\link{seshinfo_log}}, \code{\link{setup}},

[…]
