### 2018-09-13 6698e0a
updated conversion factors
