### 2018-09-13 7cc1a36
fixed minor errors, updating documentation
