+++ beavers/session_info.txt
@@ -1,0 +1,20 @@
+R version 3.4.4 (2018-03-15)
+Platform: x86_64-apple-darwin15.6.0 (64-bit)
+Running under: macOS High Sierra 10.13.6
+
+Matrix products: default
+BLAS: /System/Library/Frameworks/Accelerate.framework/Versions/A/Frameworks/vecLib.framework/Versions/A/libBLAS.dylib
+LAPACK: /Library/Frameworks/R.framework/Versions/3.4/Resources/lib/libRlapack.dylib
+
+locale:
+[1] en_GB.UTF-8/en_GB.UTF-8/en_GB.UTF-8/C/en_GB.UTF-8/en_GB.UTF-8
+
+attached base packages:
+[1] stats     graphics  grDevices utils     datasets  methods   base     
+
+other attached packages:
+[1] phylotaR_1.0.0 restez_0.1.0  
+
+loaded via a namespace (and not attached):
+ [1] compiler_3.4.4  tools_3.4.4     withr_2.1.2     treeman_1.1.2   rstudioapi_0.7 
+ [6] yaml_2.1.18     memoise_1.1.0   digest_0.6.16   sys_1.6         devtools_1.13.5
