+++ other/dev.R
@@ -3,14 +3,15 @@
 
 
 # TODO:
-# - update 'records' data with difficult records
-# - update file size calulcations
+
 
 
 devtools::load_all('.')
 restez_path_set('.')
 db_delete(everything = TRUE)
 restez_path_set('.')
+db_download(preselection = 'demo')
+
 restez_connect()
 
 restez_disconnect()
