+++ README.md
@@ -1,6 +1,7 @@
 
 <!-- README.md is generated from README.Rmd. Please edit that file -->
 <!-- devtools::rmarkdown::render("README.Rmd") -->
+<!-- Rscript -e "library(knitr); knit('README.Rmd')" -->
 Locally query GenBank <img src="logo.png" height="200" align="right"/>
 ======================================================================
 
@@ -39,53 +40,196 @@ Quick Examples
 ### Setup
 
 ``` r
+# Warning: running these examples may take a few minutes
 library(restez)
-# choose a location to store GB files
-restez_path_set('.')
-# run download function
+#> -------------
+#> restez v0.1.0
+#> -------------
+#> Remember to restez_path_set() and, then, restez_connect()
+# choose a location to store GenBank files
+restez_path_set(tempdir())
+#> ... Creating '/var/folders/ps/g89999v12490dmp0jnsfmykm0043m3/T//RtmpdUEnD6/restez'
+#> ... Creating '/var/folders/ps/g89999v12490dmp0jnsfmykm0043m3/T//RtmpdUEnD6/restez/downloads'
+# Running the download function ....
 # interactively choose GenBank files to download
-gb_download()
-# create local database
+# e.g. 20 is smallest
+db_download(preselection = '20')
+#> ───────────────────────────────────────────────────────────────────────────────────────────────
+#> Looking up latest GenBank release ...
+#> ... release number 227
+#> ... found 3164 sequence files
+#> ───────────────────────────────────────────────────────────────────────────────────────────────
+#> Which sequence file types would you like to download?
+#> Choose from those listed below:
+#> ● 1  -  'Bacterial' (520 files and 118 GB)
+#> ● 2  -  'EST (expressed sequence tag)' (485 files and 242 GB)
+#> ● 3  -  'Constructed' (369 files and 84 GB)
+#> ● 4  -  'Patent' (337 files and 77.3 GB)
+#> ● 5  -  'GSS (genome survey sequence)' (308 files and 117 GB)
+#> ● 6  -  'TSA (transcriptome shotgun assembly)' (234 files and 53.9 GB)
+#> ● 7  -  'Plant sequence entries (including fungi and algae),' (198 files and 42.6 GB)
+#> ● 8  -  'HTGS (high throughput genomic sequencing)' (155 files and 36.7 GB)
+#> ● 9  -  'Invertebrate' (108 files and 23.8 GB)
+#> ● 10  -  'Environmental sampling' (103 files and 23.9 GB)
+#> ● 11  -  'Other vertebrate' (94 files and 19.3 GB)
+#> ● 12  -  'Primate' (59 files and 13.7 GB)
+#> ● 13  -  'Viral' (57 files and 12.9 GB)
+#> ● 14  -  'Other mammalian' (55 files and 9.41 GB)
[…]
