+++ beavers/log.txt
@@ -1,53 +1,0 @@
--------------------------------------------------
-phylotaR: Implementation of PhyLoTa in R [v1.0.0]
--------------------------------------------------
-Checking for valid NCBI BLAST+ Tools ...
-Found: [/usr/local/ncbi/blast/bin/makeblastdb]
-Found: [/usr/local/ncbi/blast/bin/blastn]
-Setting up pipeline with the following parameters:
-. blstn      [/usr/local/ncbi/blast/bin/blastn]
-. btchsz     [100]
-. date       [2018-09-13]
-. mdlthrs    [3000]
-. mkblstdb   [/usr/local/ncbi/blast/bin/makeblastdb]
-. mncvrg     [51]
-. mnsql      [250]
-. mxevl      [1e-10]
-. mxnds      [1e+05]
-. mxrtry     [100]
-. mxsql      [2000]
-. mxsqs      [50000]
-. ncps       [1]
-. txid       [1963757]
-. v          [TRUE]
-. wd         [beavers]
--------------------------------------------------
---------------------------------------------
-Starting stage TAXISE: [2018-09-13 14:39:41]
---------------------------------------------
-Searching taxonomic IDs ...
-Downloading taxonomic records ...
-. [1-100]
-. [101-200]
-. [201-300]
-. [301-331]
-Generating taxonomic dictionary ...
----------------------------------------------
-Completed stage TAXISE: [2018-09-13 14:40:02]
----------------------------------------------
-----------------------------------------------
-Starting stage DOWNLOAD: [2018-09-13 14:40:02]
-----------------------------------------------
-Identifying suitable clades ...
-Identified [1] suitable clades.
-Downloading hierarchically ...
-Working on parent [id 1963757]: [1/1] ...
-. + direct ...
-. + by child ...
-. . Working on child [id 29132]
-. . + direct ...
-. . + by child ...
-. . . Working on child [id 10184]
-. . . + direct ...
-. . . . Downloading [7 sqs] ...
-. . . . . [1-7]
