+++ README.md
@@ -20,7 +20,7 @@ This package aims to make sequence retrieval more efficient by allowing a user t
 
 <img src="https://raw.githubusercontent.com/AntonelliLab/restez/master/paper/outline.png" height="500" align="center"/>
 
-**For more detailed information on the pacakge's functions and detailed guides on downloading, constructing and querying a database, visit the [restez website](https://antonellilab.github.io/restez/index.html).**
+**For more detailed information on the package's functions and detailed guides on downloading, constructing and querying a database, visit the [restez website](https://antonellilab.github.io/restez/index.html).**
 
 Installation
 ------------
@@ -48,11 +48,11 @@ library(restez)
 #> Remember to restez_path_set() and, then, restez_connect()
 # choose a location to store GenBank files
 restez_path_set(tempdir())
-#> ... Creating '/var/folders/ps/g89999v12490dmp0jnsfmykm0043m3/T//RtmpdUEnD6/restez'
-#> ... Creating '/var/folders/ps/g89999v12490dmp0jnsfmykm0043m3/T//RtmpdUEnD6/restez/downloads'
+#> ... Creating '/var/folders/ps/g89999v12490dmp0jnsfmykm0043m3/T//Rtmpk2tcxZ/restez'
+#> ... Creating '/var/folders/ps/g89999v12490dmp0jnsfmykm0043m3/T//Rtmpk2tcxZ/restez/downloads'
 # Running the download function ....
 # interactively choose GenBank files to download
-# e.g. 20 is smallest
+# e.g. Unannotated sequences, 20, is the smallest
 db_download(preselection = '20')
 #> ───────────────────────────────────────────────────────────────────────────────────────────────
 #> Looking up latest GenBank release ...
@@ -122,11 +122,11 @@ id <- sample(list_db_ids(), 1)
 # sequences
 seq <- gb_sequence_get(id)[[1]]
 print(seq)
[…]
