### 2018-09-13 b95c999
updated printed info
