+++ other/dev.R
@@ -10,7 +10,7 @@ devtools::load_all('.')
 restez_path_set('.')
 db_delete(everything = TRUE)
 restez_path_set('.')
-db_download(preselection = 'demo')
+db_download()
 
 restez_connect()
 
