+++ README.md
@@ -48,8 +48,8 @@ library(restez)
 #> Remember to restez_path_set() and, then, restez_connect()
 # choose a location to store GenBank files
 restez_path_set(tempdir())
-#> ... Creating '/var/folders/ps/g89999v12490dmp0jnsfmykm0043m3/T//Rtmpk2tcxZ/restez'
-#> ... Creating '/var/folders/ps/g89999v12490dmp0jnsfmykm0043m3/T//Rtmpk2tcxZ/restez/downloads'
+#> ... Creating '/var/folders/ps/g89999v12490dmp0jnsfmykm0043m3/T//RtmpcuQynQ/restez'
+#> ... Creating '/var/folders/ps/g89999v12490dmp0jnsfmykm0043m3/T//RtmpcuQynQ/restez/downloads'
 # Running the download function ....
 # interactively choose GenBank files to download
 # e.g. Unannotated sequences, 20, is the smallest
@@ -61,31 +61,52 @@ db_download(preselection = '20')
 #> ───────────────────────────────────────────────────────────────────────────────────────────────
 #> Which sequence file types would you like to download?
 #> Choose from those listed below:
-#> ● 1  -  'Bacterial' (520 files and 118 GB)
-#> ● 2  -  'EST (expressed sequence tag)' (485 files and 242 GB)
-#> ● 3  -  'Constructed' (369 files and 84 GB)
-#> ● 4  -  'Patent' (337 files and 77.3 GB)
-#> ● 5  -  'GSS (genome survey sequence)' (308 files and 117 GB)
-#> ● 6  -  'TSA (transcriptome shotgun assembly)' (234 files and 53.9 GB)
-#> ● 7  -  'Plant sequence entries (including fungi and algae),' (198 files and 42.6 GB)
-#> ● 8  -  'HTGS (high throughput genomic sequencing)' (155 files and 36.7 GB)
-#> ● 9  -  'Invertebrate' (108 files and 23.8 GB)
-#> ● 10  -  'Environmental sampling' (103 files and 23.9 GB)
-#> ● 11  -  'Other vertebrate' (94 files and 19.3 GB)
-#> ● 12  -  'Primate' (59 files and 13.7 GB)
-#> ● 13  -  'Viral' (57 files and 12.9 GB)
-#> ● 14  -  'Other mammalian' (55 files and 9.41 GB)
-#> ● 15  -  'Rodent' (31 files and 7.3 GB)
-#> ● 16  -  'STS (sequence tagged site)' (20 files and 4.45 GB)
-#> ● 17  -  'HTC (high throughput cDNA sequencing)' (15 files and 3.42 GB)
-#> ● 18  -  'Synthetic and chimeric' (10 files and 2.41 GB)
-#> ● 19  -  'Phage' (5 files and 1.06 GB)
-#> ● 20  -  'Unannotated' (1 files and 0.00108 GB)
+#> ● 1  - 'Bacterial'
+#>         520 files and 118 GB
+#> ● 2  - 'EST (expressed sequence tag)'
+#>         485 files and 242 GB
+#> ● 3  - 'Constructed'
+#>         369 files and 84 GB
+#> ● 4  - 'Patent'
+#>         337 files and 77.3 GB
+#> ● 5  - 'GSS (genome survey sequence)'
+#>         308 files and 117 GB
[…]
