### 2018-09-13 824be16
minor improvements
