### 2018-09-14 0a40d5a
added sequence count
