+++ other/dev.R
@@ -13,7 +13,12 @@ restez_path_set('.')
 db_download()
 
 restez_connect()
-
+db_create(min_length = 10, max_length = 1000)
+restez_status()
+db_delete()
+restez_connect()
+db_create()
+restez_status()
 restez_disconnect()
 
 
