+++ examples/restez_disconnect.R
@@ -6,6 +6,6 @@ demo_db_create()
 restez_status()
 # always remember to disconnect from a database when you've finished
 restez_disconnect()
-db_delete()
+db_delete(everything = TRUE)
 # Errors
 # restez_status()
