+++ examples/demo_db_create.R
@@ -8,5 +8,4 @@ demo_db_create(n = 100)
 (gb_sequence_get(id = 'demo_1'))
 
 # Delete a demo database after an example
-restez_disconnect()
-db_delete()
+db_delete(everything = TRUE)
