+++ examples/count_db_ids.R
@@ -2,9 +2,7 @@ library(restez)
 restez_path_set(filepath = tempdir())
 restez_connect()
 demo_db_create()
-(ver <- gb_version_get(id = 'demo_1'))
-(vers <- gb_version_get(id = c('demo_1', 'demo_2')))
-
+(count_db_ids())
 
 # delete demo after example
 db_delete(everything = TRUE)
