### 2018-09-14 440ac3b
future proofing genbank download details
