### 2018-09-14 4b7cdd2
better explanation of seq types
