+++ R/rentrez-wrappers.R
@@ -6,6 +6,9 @@
 #' parameters, if the record is missing in the database, the function then
 #' calls rentrez::entrez_fetch to search GenBank remotely.
 #' 
+#' \code{rettype='fasta'} and \code{rettype='gb'} are respectively equivalent to 
+#' \code{\link{gb_fasta_get}} and \code{\link{gb_record_get}}.
+#' 
 #' @note It is advisable to call restez and rentrez functions with '::' notation
 #' rather than library() calls to avoid namespace issues. e.g.
 #' restez::entrez_fetch().
