+++ man/list_db_ids.Rd
@@ -54,3 +54,4 @@ Other get: \code{\link{count_db_ids}},
   \code{\link{gb_sequence_get}},
   \code{\link{gb_version_get}}, \code{\link{is_in_db}}
 }
+\concept{get}
