+++ man/gb_df_generate.Rd
@@ -81,3 +81,4 @@ Other private: \code{\link{add_rcrd_log}},
   \code{\link{sql_path_get}}, \code{\link{status_class}},
   \code{\link{stat}}, \code{\link{testdatadir_get}}
 }
+\concept{private}
