+++ man/gb_version_get.Rd
@@ -37,3 +37,4 @@ Other get: \code{\link{count_db_ids}},
   \code{\link{gb_sequence_get}}, \code{\link{is_in_db}},
   \code{\link{list_db_ids}}
 }
+\concept{get}
