+++ man/gb_organism_get.Rd
@@ -36,3 +36,4 @@ Other get: \code{\link{count_db_ids}},
   \code{\link{gb_version_get}}, \code{\link{is_in_db}},
   \code{\link{list_db_ids}}
 }
+\concept{get}
