+++ vignettes/restez.Rmd
@@ -12,156 +12,113 @@ knitr::opts_chunk$set(
   collapse = TRUE,
   comment = "#>"
 )
+# hide line spinner
+assignInNamespace(x = 'custom_download2', value = restez:::custom_download,
+                  ns = 'restez')
 ```
 
-#0. Contents
+#Contents
  
-1. **[Download GenBank](#download-genbank)**
-2. **[Query your local database](#query)**
-3. **[Use Entrez tools](#entrez)**
+1. [Download GenBank](#download-genbank)
+2. [Query your local database](#query)
+3. [Use Entrez tools](#entrez)
 
 #1. Download GenBank
 
 In the first section of this introduction to `restez` we will explain how to download parts of GenBank on to your computer and from these downloads, construct a local database. GenBank can be downloaded from NCBI using the File Transfer Protocol (FTP), the online portal for these downloads can be viewed from your web browser,
 [FTP NCBI GenBank](ftp://ftp.ncbi.nlm.nih.gov/genbank/). GenBank is hosted at this portal as 1000s of compressed sequence files called flatfiles. New versions of these flatfiles are released once a month as new sequence information is uploaded. For information on the latest GenBank release, see the [NCBI statistics page](https://www.ncbi.nlm.nih.gov/genbank/statistics/).
 
-`restez` downloads a selection of these flatfiles, decompresses them and unpacks them into an SQLite database that can be searched using its functions. Because there are potentially huge amounts of sequence information, all of which is unlikely to be of interest to a user, `restez` allows you to select the sets of flatfiles to download. For the most part these sets of flatfile types are determined by taxonomy: a user can download all primates or plant sequences, for example.
+`restez` downloads a selection of these flatfiles and unpacks them into an SQL-like database (specifically, [MonetDB](https://www.monetdb.org/)) that can be queried using `restez`'s functions. Because there are potentially huge amounts of sequence information, all of which is unlikely to be of interest to a user, `restez` allows you to select the sets of flatfiles to download based on GenBank's domains. For the most part these sets of flatfile types are determined by taxonomy, for example, a user can download all primate or plant sequences.
 
 Here, we will explain how to set up `restez`, download genbank and then create a database from the downloaded files.
 
 ##1.1 Setting up
 
-Before we can download, we need to tell `restez` where we would like to store the downloaded files and the eventual GenBank database. To do this we use the `restez_path_set` function. A user provides a system file path to a folder, and within that folder a new folder called 'restez' is created. In this example, I will create a `restez` folder in my desktop.
+Before we can download anything, we need to tell `restez` where we would like to store the downloaded files and the eventual GenBank database. To do this we use the `restez_path_set` function. With this function a user can provide a system file path to a folder within which the function will create a new folder called 'restez'. In this example, we will create a `restez` folder in a temporary folder.
 
-```{r setting up, eval=FALSE, include=TRUE}
+```{r presetup, include=FALSE}
+rstz_pth <- tempdir()
+restez::restez_path_set(filepath = rstz_pth)
+restez::db_delete(everything = TRUE)
+```
+```{r setting up, echo=TRUE}
 library(restez)
-# On my linux computer and I specify my desktop
-fp <- '~/Desktop'
-restez_path_set(filepath = fp)
-```
-```
-... Creating '~/Desktop/restez'
-... Creating '~/Desktop/restez/downloads'
+# rstz_pth <- tempdir()
+restez_path_set(filepath = rstz_pth)
 ```
 
-Messages are printed to console telling us that the path was created and that a downloads subfolder was created.
[…]
