+++ README.Rmd
@@ -21,7 +21,9 @@ assignInNamespace(x = 'custom_download2', value = restez:::custom_download,
 
 [![Build Status](https://travis-ci.org/AntonelliLab/restez.svg?branch=master)](https://travis-ci.org/AntonelliLab/restez) [![Coverage Status](https://coveralls.io/repos/github/AntonelliLab/restez/badge.svg?branch=master)](https://coveralls.io/github/AntonelliLab/restez?branch=master) [![DOI](https://zenodo.org/badge/129107980.svg)](https://zenodo.org/badge/latestdoi/129107980)
 
-Download parts of [NCBI's GenBank](https://www.ncbi.nlm.nih.gov/nuccore) to a local folder and create a simple SQLite database. Use 'get' tools to query the database by accession IDs. [rentrez](https://github.com/ropensci/rentrez) wrappers are available, so that if sequences are not available locally they can be searched for online through [Entrez](https://www.ncbi.nlm.nih.gov/books/NBK25500/). Visit the [website](https://antonellilab.github.io/restez/index.html) to find out more.
+Download parts of [NCBI's GenBank](https://www.ncbi.nlm.nih.gov/nuccore) to a local folder and create a simple SQL-like database. Use 'get' tools to query the database by accession IDs. [rentrez](https://github.com/ropensci/rentrez) wrappers are available, so that if sequences are not available locally they can be searched for online through [Entrez](https://www.ncbi.nlm.nih.gov/books/NBK25500/).
+
+See the [detailed  tutorials](https://antonellilab.github.io/restez/articles/restez.html) for more information.
 
 ## Introduction
 
@@ -33,32 +35,40 @@ This package aims to make sequence retrieval more efficient by allowing a user t
 
 <img src="https://raw.githubusercontent.com/AntonelliLab/restez/master/paper/outline.png" height="500" align="center"/>
 
-**For more detailed information on the package's functions and detailed guides on downloading, constructing and querying a database, visit the [restez website](https://antonellilab.github.io/restez/index.html).**
-
 ## Installation
 
-You can install restez from GitHub with:
+You can install `restez` from GitHub with:
 
-```{r gh-installation, eval = FALSE}
+```{r gh-installation, include=TRUE, echo=TRUE, eval=FALSE}
 # install.packages("devtools")
 devtools::install_github("AntonelliLab/restez")
 ```
 
 ## Quick Examples
 
-> For more detailed tutorials, visit the [restez website](https://antonellilab.github.io/restez/index.html).
+> For more detailed information on the package's functions and detailed guides on downloading, constructing and querying a database, see the [detailed  tutorials](https://antonellilab.github.io/restez/articles/restez.html).
 
 ### Setup
 
-```{r gb-download, echo=TRUE}
+```{r presetup, include=FALSE}
+rstz_pth <- tempdir()
+restez::restez_path_set(filepath = rstz_pth)
+restez::db_delete(everything = TRUE)
+```
+```{r restez-setup, echo=TRUE, eval=TRUE, results='hide'}
 # Warning: running these examples may take a few minutes
 library(restez)
 # choose a location to store GenBank files
-restez_path_set(tempdir())
-# Running the download function ....
-# interactively choose GenBank files to download
-# e.g. Unannotated sequences, 20, is the smallest
-db_download(preselection = '20')
+restez_path_set(rstz_pth)
+```
+``` {r reassigndownload, include=FALSE, eval=TRUE}
+db_download <- function() {
[…]
