+++ vignettes/search_and_fetch.Rmd
@@ -3,7 +3,7 @@ title: "2. How to search for and fetch sequences"
 date: "`r Sys.Date()`"
 output: rmarkdown::html_vignette
 vignette: >
-  %\VignetteIndexEntry{Search and fetch}
+  %\VignetteIndexEntry{2. How to search for and fetch sequences}
   %\VignetteEngine{knitr::rmarkdown}
   %\VignetteEncoding{UTF-8}
 ---
@@ -15,33 +15,16 @@ knitr::opts_chunk$set(
 )
 ```
 
-In this tutorial we are going to search NCBI with a custom search term, download accession IDs, and retrieve fasta sequences using `restez`.
+A downside of the `restez` approach is we are unable to then search the local database for sequences of interest as the database cannot possibly contain sufficient metadata (most notably taxonomic information) to perform as rich a search as online via NCBI. As a result we must perform the sequence discovery process independently from `restez`. In this tutorial we will demonstrate how to peform online searches using the `rentrez` package and then use the search results to look up sequences in a `restez` database.
 
-## Setup
+To run all the code in this tutorial you will need to have already set up the rodents database, see [Build a database of all rodents](https://antonellilab.github.io/restez/articles/rodents.html).
 
-First, we need to download and setup the restez database. When we use `gb_download()` we are going to select 'rodents'. We are downloading rodents because they are the smallest available set of sequences, minismising the download time for this demonstration. If you have already downloaded the rodents flatfiles as explained in previous tutorials, then simply set your restez path to wherever you previously set your restez path.
+#Search NCBI
 
-```{r setup-database, eval=FALSE, include=TRUE}
-library(restez)
-fp <- file.path(getwd(), 'demo')
-dir.create(fp)
-restez_path_set(fp)
-gb_download()
-db_create()
-```
-
-## Search NCBI
-
-After setting up the database, we can then identify the sequences we are interested in. For this, we cannot use restez to perform a search,
-because the download just doesn't contain enough metadata for the full range of possible search arguments.
-
-Let's pretend we're intersted in COI sequences for a group of Rodents called the Sciuromorpha. We can create an NCBI search term,
-use `entrez_search` to perform the search and then retrieve the accession IDs using `entrez_fetch` and the `acc` rettype. `entrez_fetch`
-returns a single text that will need to be split up by the newline character. Finally, we should drop any version number after the downloaded
-accessions. This is not strictly necessary, but it makes it easier to check our results from restez later.
+Let's pretend we're interested in COI sequences for a group of rodents called the Sciuromorpha. We can create an NCBI search term; use `entrez_search` to perform the search; and then retrieve the accession IDs using `entrez_fetch` with the `acc` rettype. `entrez_fetch` returns a single text that will need to be split up by the newline character. Finally, we should drop any version number after the downloaded accessions. This is not strictly necessary, but it makes it easier to check our results from `restez` later.
 
 
-```{r search, eval=FALSE, include=TRUE}
+```{r search, include=TRUE}
 # 33553 - Sciuromorpha - squirrel-like things
 search_term <- 'txid33553[Organism:exp] AND COI [GENE]'
 search_object <- rentrez::entrez_search(db = 'nucleotide', term = search_term,
@@ -51,44 +34,57 @@ accessions <- rentrez::entrez_fetch(db = 'nucleotide',
                                     rettype = 'acc')
 accessions <- strsplit(x = accessions, split = '\\n')[[1]]
 accessions <- sub(pattern = '\\.[0-9]+', replacement = '', x = accessions)
+print(accessions[1:10])
 ```
 
+#Retrieve sequences
 
-## Retrieve sequences
[…]
