+++ vignettes/parsing.Rmd
@@ -2,7 +2,7 @@
 title: "3. Advanced parsing of a GenBank record"
 output: rmarkdown::html_vignette
 vignette: >
-  %\VignetteIndexEntry{parsing}
+  %\VignetteIndexEntry{3. Advanced parsing of a GenBank record}
   %\VignetteEngine{knitr::rmarkdown}
   %\VignetteEncoding{UTF-8}
 ---
@@ -14,13 +14,11 @@ knitr::opts_chunk$set(
 )
 ```
 
-#Introduction
-
-In this vigennette we will demonstrate how the `gb_extract()` function works. Restez downloads and stores all GenBank records in text format. Ordinarily, to be able to extract relevant bits of information from a text record in a systematic way we would need to make use of [regular expressions](https://en.wikipedia.org/wiki/Regular_expression). `gb_extract()` uses regular expressions, so we don't have to. Here's how it works.
+In this tutorial we will demonstrate how the `gb_extract()` function works. `restez` downloads and stores all GenBank records in text format. Ordinarily, to be able to extract relevant bits of information from a text record in a systematic way we would need to make use of [regular expressions](https://en.wikipedia.org/wiki/Regular_expression). `gb_extract()` uses regular expressions, so we don't have to. Here's how it works.
 
 #A GenBank record
 
-restez comes with an example GenBank record, AY952423, which can be viewed [online](https://www.ncbi.nlm.nih.gov/nuccore/AY952423.1). We can retrieve the record with `rentrez` or load it using the `data()`. We can visual the record as it appears online using `cat()` -- like print but parses the newline spaces correctly.
+restez comes with an example GenBank record, AY952423, which can be viewed [online](https://www.ncbi.nlm.nih.gov/nuccore/AY952423.1). We can retrieve the record with `rentrez` or load it using the `data()`. We can visualise the record in R as it appears online using `cat()` -- like print but with newline spaces parsed correctly.
 
 ```{r record, echo=TRUE}
 library(restez)
@@ -49,17 +47,28 @@ We can extract different elements of the above record with `gb_extract()` ....
 # even the features as a list object
 features <- gb_extract(record = record, what = 'features')
 print(features[[1]])
-# and of coure the sequence itself
+# and of course the sequence itself
 seq <- gb_extract(record = record, what = 'sequence')
 str(seq)
 ```
 
 #From the database
 
-You can try and the above functions out manually by downloading them through the [`rentrez` package](https://github.com/ropensci/rentrez). You can also parse the text formatted records from a restez database.
+You can try out the above functions yourself on any sequence record by downloading them through the [`rentrez` package](https://github.com/ropensci/rentrez) using `entrez_fetch(db='nucelotide', rettype='gb')`. Or why not test them out using any of the records from the rodents database?
 
-```{r database, eval=FALSE, include=TRUE}
-# ensure restez database is already set-up
-record <- gb_record_get(id = 'ACCESSIONID')
-gb_extract(record = record, what = 'features')
+```{r pathset, include=FALSE}
+rodents_path <- file.path(getwd(), 'rodents')
 ```
+```{r database, include=TRUE}
+library(restez)
+restez_path_set(rodents_path)
+restez_connect()
+(rand_id <- sample(suppressWarnings(list_db_ids()), 1))
+record <- gb_record_get(rand_id)
+(gb_extract(record = record, what = 'features'))
+restez_disconnect()
+```
+
+#Next up
+
+**[Running phylotaR with restez](https://antonellilab.github.io/restez/articles/phylotar.html)**
