+++ vignettes/parsing.Rmd
@@ -16,7 +16,7 @@ knitr::opts_chunk$set(
 
 In this tutorial we will demonstrate how the `gb_extract()` function works. `restez` downloads and stores all GenBank records in text format. Ordinarily, to be able to extract relevant bits of information from a text record in a systematic way we would need to make use of [regular expressions](https://en.wikipedia.org/wiki/Regular_expression). `gb_extract()` uses regular expressions, so we don't have to. Here's how it works.
 
-#A GenBank record
+##A GenBank record
 
 restez comes with an example GenBank record, AY952423, which can be viewed [online](https://www.ncbi.nlm.nih.gov/nuccore/AY952423.1). We can retrieve the record with `rentrez` or load it using the `data()`. We can visualise the record in R as it appears online using `cat()` -- like print but with newline spaces parsed correctly.
 
@@ -27,7 +27,7 @@ data(record)
 cat(record)
 ```
 
-#Extracting
+##Extracting
 
 We can extract different elements of the above record with `gb_extract()` ....
 
@@ -52,12 +52,13 @@ seq <- gb_extract(record = record, what = 'sequence')
 str(seq)
 ```
 
-#From the database
+##From the database
 
 You can try out the above functions yourself on any sequence record by downloading them through the [`rentrez` package](https://github.com/ropensci/rentrez) using `entrez_fetch(db='nucelotide', rettype='gb')`. Or why not test them out using any of the records from the rodents database?
 
 ```{r pathset, include=FALSE}
-rodents_path <- file.path(getwd(), 'rodents')
+pkgwd <- sub(pattern = 'vignettes', replacement = '' , x = getwd())
+rodents_path <- file.path(pkgwd, 'rodents')
 ```
 ```{r database, include=TRUE}
 library(restez)
@@ -69,6 +70,6 @@ record <- gb_record_get(rand_id)
 restez_disconnect()
 ```
 
-#Next up
+##Next up
 
 **[Running phylotaR with restez](https://antonellilab.github.io/restez/articles/phylotar.html)**
