+++ vignettes/phylotar.Rmd
@@ -14,72 +14,67 @@ knitr::opts_chunk$set(
 )
 ```
 
-#Intoduction
+In this tutorial we will showcase how a `restez` database can be used to speed up a [phylotaR](https://github.com/AntonelliLab/phylotaR) run. `phylotaR` runs an automated pipeline for identifying ortholgous gene clusters as the first step in a phylogenetic analysis. A user provides a taxonomic identity and the pipeline downloads all relevant sequences and identifies clusters using a local-alignment search tool. For more information on `phylotaR` see its [published article](https://doi.org/10.3390/life8020020).
 
-In this demonstration we will showcase how a `restez` database can be used to speed up a [phylotaR](https://github.com/AntonelliLab/phylotaR) run. phylotaR runs an automated pipeline for identifying ortholgous gene clusters as the first step in a phylogenetic analysis. A user provides a taxonomic identity and the pipeline downloads all relevant sequences and identifies clusters using a local-alignment search tool. For more information on phylotaR see its [published article](https://doi.org/10.3390/life8020020).
+By using `restez` in conjunction with `phylotaR`, we will not only being saving time, but also improving the chances of a successful `phylotaR` run -- often NCBI Entrez limits the number of requests or even rejects requests from IP addresses that are making too many. Note, however, that the gains in using `restez` with `phylotaR` only make sense if you make use of the `restez` database multiple times or if you wish to radically increase the maximum number of sequences to download per taxon (by default it is only 3,000). Also, note that using a `restez` database does not currently eliminate the need for an internet connection. `phylotaR` still needs to look up taxonomic information and must also identify relevant sequence IDs using Entrez (this may change in the future as `restez` develops).
 
-By using restez in conjunction with phylotaR, we will not only being saving time, but also improving the chances of a successful phylotaR run -- often NCBI Entrez limits the number of requests or even rejects requests from IP addresses that are making too many. Note, however, that the gains in using restez with phylotaR only make sense if you make use of the restez database multiple times or if you wish to radically increase the maximum number of sequences to download per taxon (by default it is only 3,000). Also, note that using a restez database does not currently eliminate the need for an internet connection. phylotaR still needs to look up taxonomic information and must also identify relevant sequence IDs using Entrez (this may change in the future as restez develops).
+We will run a `phylotaR` run for the rodent genus, *Cavia*, the clade that contains guinea pigs. Because this clade is so small, it should not take too long to run it. We will, however, limit to showcasing only the download stage of the `phylotaR` pipeline, where `restez` is required.
 
-We will run a phylotaR run for the beaver-like, rodent suborder, Caviomorpha, the clade that contains things like beavers and gophers. We will first set-up our restez database followed by our phylotaR pipeline and then run the pipeline. To run the entire pipeline can take a long time (1 or 2 hours, with most time being taken by the database set-up) but depends on your internet connection and computer specs.
-
-#Restez set-up
-
-First we need to create a local GenBank database. Because the Caviomorpha are rodents we must select the rodent GenBank files.
[…]

[…]
