### 2018-09-16 07670ac
reworked function families
