+++ man/db_download.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/db-tools.R
+% Please edit documentation in R/db-setup-tools.R
 \name{db_download}
 \alias{db_download}
 \title{Download database}
