+++ man/list_db_ids.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/general-get-tools.R
+% Please edit documentation in R/db-get-tools.R
 \name{list_db_ids}
 \alias{list_db_ids}
 \title{List database IDs}
@@ -46,12 +46,7 @@ cat(fasta_file)
 db_delete(everything = TRUE)
 }
 \seealso{
-Other get: \code{\link{count_db_ids}},
-  \code{\link{gb_definition_get}},
-  \code{\link{gb_fasta_get}},
-  \code{\link{gb_organism_get}},
-  \code{\link{gb_record_get}},
-  \code{\link{gb_sequence_get}},
-  \code{\link{gb_version_get}}, \code{\link{is_in_db}}
+Other db: \code{\link{count_db_ids}},
+  \code{\link{is_in_db}}
 }
-\concept{get}
+\concept{db}
