+++ man/count_db_ids.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/general-get-tools.R
+% Please edit documentation in R/db-get-tools.R
 \name{count_db_ids}
 \alias{count_db_ids}
 \title{Return the number of ids}
@@ -29,12 +29,6 @@ demo_db_create()
 db_delete(everything = TRUE)
 }
 \seealso{
-Other get: \code{\link{gb_definition_get}},
-  \code{\link{gb_fasta_get}},
-  \code{\link{gb_organism_get}},
-  \code{\link{gb_record_get}},
-  \code{\link{gb_sequence_get}},
-  \code{\link{gb_version_get}}, \code{\link{is_in_db}},
-  \code{\link{list_db_ids}}
+Other db: \code{\link{is_in_db}}, \code{\link{list_db_ids}}
 }
-\concept{get}
+\concept{db}
