+++ man/demo_db_create.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/db-tools.R
+% Please edit documentation in R/db-setup-tools.R
 \name{demo_db_create}
 \alias{demo_db_create}
 \title{Create demo database}
