### 2018-09-16 e6655b5
do not want to build vignettes in package
