### 2018-09-16 a5269f3
missing dep for tests R.utils
