+++ DESCRIPTION
@@ -28,6 +28,7 @@ Imports:
 Suggests:
     testthat,
     knitr,
+    R.utils,
     rmarkdown
 RoxygenNote: 6.1.0
 VignetteBuilder: knitr
