### 2018-09-16 b091a11
final updates before review
