+++ man/restez_path_get.Rd
@@ -21,10 +21,7 @@ restez_path_set(filepath = tempdir())
 (restez_path_get())
 }
 \seealso{
-Other setup: \code{\link{db_create}},
-  \code{\link{db_delete}}, \code{\link{db_download}},
-  \code{\link{demo_db_create}},
-  \code{\link{restez_connect}},
+Other setup: \code{\link{restez_connect}},
   \code{\link{restez_disconnect}},
   \code{\link{restez_path_set}},
   \code{\link{restez_path_unset}},
