+++ man/count_db_ids.Rd
@@ -29,6 +29,9 @@ demo_db_create()
 db_delete(everything = TRUE)
 }
 \seealso{
-Other db: \code{\link{is_in_db}}, \code{\link{list_db_ids}}
+Other database: \code{\link{db_create}},
+  \code{\link{db_delete}}, \code{\link{db_download}},
+  \code{\link{demo_db_create}}, \code{\link{is_in_db}},
+  \code{\link{list_db_ids}}
 }
-\concept{db}
+\concept{database}
