+++ R/db-setup-tools.R
@@ -123,7 +123,7 @@ db_download <- function(db='nucleotide', overwrite=FALSE, preselection=NULL) {
 #' all environmental sequences twice, you can generate multiple restez databases
 #' using the same downloaded files from a single restez path.
 #'
-#' This function will not overwrite a pre-exisitng database. Old databases must
+#' This function will not overwrite a pre-existing database. Old databases must
 #' be deleted before a new one can be created. Use \code{\link{db_delete}} with
 #' everything=FALSE to delete an SQL database.
 #'
