+++ R/log-tools.R
@@ -38,7 +38,7 @@ seshinfo_log <- function() {
 
 #' @name db_sqlngths_log
 #' @title Log the min and max sequence lengths
-#' @description Log the min and maximum sequnece length used in the created db.
+#' @description Log the min and maximum sequence length used in the created db.
 #' @param min_lngth Minimum length
 #' @param max_lngth Maximum length
 #' @return NULL
@@ -89,7 +89,7 @@ filename_log <- function(fl, fp) {
 #' @description This function is called whenever a file is successfully
 #' downloaded. A row entry is added to the 'download_log.tsv' in the user's
 #' restez path containing the file name, the GB release number and the time of
-#' successfuly download. The log is to help users keep track of when they
+#' successfully download. The log is to help users keep track of when they
 #' downloaded files and to determine if the downloaded files are out of date.
 #' @param fl file name, character
 #' @return NULL
