+++ other/spell_check.R
@@ -20,3 +20,9 @@ ignore <- c('restez',
             'SQLite')
             #'uilist')
 devtools::spell_check(ignore = ignore, dict = 'en_GB')
+vignette_files <- file.path(getwd(), 'vignettes',
+                            list.files('vignettes', pattern = 'Rmd'))
+for (fl in vignette_files) {
+  print(fl)
+  print(devtools:::spell_check_file(fl, ignore = ignore, dict = 'en_GB'))
+}
