+++ R/download-tools.R
@@ -22,7 +22,7 @@ predict_datasizes <- function(uncompressed_filesize) {
   caveat <- paste0('\nPlease note, the real sizes of the database and its ',
                    'downloads cannot be\naccurately predicted beforehand.\n',
                    'These are just estimates, actual sizes may differ by up ',
-                   'to 0-5GB.\n')
+                   'to a factor of two.\n')
   cat_line(crayon::italic(caveat))
 }
 
