### 2018-09-16 #232 @DomBennett
Firstly, thanks to you both for spending so much time reviewing the package. It’s great to have such detailed feedback and useful comments. I hope my efforts at revising the package do your good reviews justice.

Firstly, in response to @eveskew’s comments.

- *Explain Maintenance of Multiple SQL Databases*: I’ve done this as part of the ‘Tips and tricks’ vignette.
- *More Informative Error Messages When restez Path Not Set*: All db interactions go through the function connection_get. It raises an error if the db is not connected. Equally, if no path is set restez_connect will also raise an informative error. Hopefully, this will reduce the number of obscure error messages.
- *Change list_db_ids() Defaults*: I now raise a warning. I am reluctant to drop the default n. I always accidentally generate huge vectors.
- *Don't Export restez_ready()?*: I think restez_ready is useful as a quick T/F for if the database can be read. I can envisage it being very useful in a programmatic context where streams of text printed to the console is not helpful. Instead, I have further developed the restez_status function so that it returns a structured list of the current setup status, not T/F.
- *Update restez_status() With Number of Records*: restez_status does this now and there is also a new function called count_db_ids which should be faster than list_db_ids. Also, restez_status does a lot more than before. I want it to be the function a user calls to help them work out what state their current setup is. It should be the go to function whenever you get confused.
- *Clarify SQL Database Size Requirements*: There was nothing wrong with your setup. It’s just hard to predict how much diskspace will be taken up. I’ve attempted to improve the process by using the stated uncompressed file sizes in the genbank release notes. To make users aware that it is just a prediction I have also added a little caveat text.
- *Revise References To Defunct Functions*: I’ve hopefully dropped all defunct function names from documentation now.
- *Differing gb_record_get() and entrez_fetch() Behavior*: In your example you’re asking for two different things: gb record and fasta. entrez_fetch(rettype=’gb’) and gb_record_get() should return the same thing. Equally, entrez_fetch(rettype=’fasta’) and gb_fasta_get() should return the same thing. I understand that it is not immediately obvious what the different sequence formats are, so I have updated the documentation in both functions to help clarify what to expect.
- *Typo Fixes -- future proofing*: I’ve done a final check for typos. I also took onboard your suggestion of future-proofing the db_download text. It now changes the example text depending on which numbers are listed as rodent, primate or other mammal.
- *Make Vignette Availability More Obvious*: I’ve done as you suggested!
- *Revise README Graphic*: I have adapted the graphic as per your suggestions. All the querying functions are now listed and the restez path has been moved to the top.
- *Better Organize Function Types (Families)*: There are now five families in keeping with `?restez`, the website and the roxygen families.
- *Harmonize Vignette Workflows*: I’ve done this by introducing an order in which to work through the tutorials. I’ve also separated the rodent database set-up from each of the tutorials and made all tutorials dependent on that database.
- *Performance Issues With Querying Large Databases*: Yeah performance was a big problem. I switched the database to MonetDB and I also switched the responsibility of connecting to the db from the internal functions to the user. Connecting to a large db can be the factor that takes the most time, so although it is extra work for the user to ensure the db is open and closed, it saves a lot of time if a user is performing lots of queries. With the new restez package I was able to create a viral db and make a query for the definition of one random id. After restez_connect() completes, the query itself can be performed in a fraction of a second.
- *roxygen2 Update*: I’ve updated roxygen on all my machines.

Secondly, in response to @naupaka’s comments:

- *JOSS paper.md*: I copied the format suggested in the JOSS documentation from this [example](https://github.com/arfon/fidgit/tree/master/paper). Is that correct @sckott?
- *Issues creating DB*: I looked into what was causing these issues and fixed them. The problem was essentially down to record formats that I did not expect (e.g. records without sequence, keywords in descriptions). To detect other such problems I ran custom download and create scripts that would select random subsamples of sequence files representing each of the GenBank domains. Doing this I was able to detect other problems with my record extraction functions and fix them. Since fixing these problems I have been able to run `restez` across any of the sequence file types hosted by GenBank without problems. In my most recent run I was able to create a database for all ESTs representing >200 GB of data. See below. I recommend not building too large a database because queries can be super slow. Particularly for databases with large sequences.
- *Specify package in README*: I only recommend using `rentrez::` in the documentation to prevent clashes with `restez`. I think introducing the `::` format in the README might be a bit much for user’s unfamiliar with it. I do, however, recommend using `restez::` in the vignette on phylotaR. Perhaps I should add additional explanation on to when is best to use `::` and `library` in the Tips and Tricks? (On an independent note, I also suggest exploring the internal functions with `:::` which may also cause confusion but is suggested only in the depths of the documentation rather than face-up.)
- *Give supported return types instead of just not supported ones*: I’ve updated this as per @eveskew’s comments.
- *Vignette output for examples out of sync with how current functions run*: I’ve corrected this by making the vignette chunks push output into the markdown text. On the downside this means generating my pkgdown site is a lot slower.
- *Estimated sizes of downloads and db don't agree with reported size after download and db creation*: As described above, I have attempted to caveat and improve these estimations.

Big thanks again to you both. Responding to your points took me a lot longer than I anticipated, highlighting the importance of good code review. Personally speaking, I don’t feel the package is quite neatly doing what I wanted it to do. So if you have any more useful ideas for improving it, I am all ears.

Thanks,
Dom

**EST database example**
```
> restez_status()
Checking setup status at  ...
────────────────────────────────────────────────────────────────────────────────
Restez path ...
... Path './restez'
... Does path exist? 'Yes'
────────────────────────────────────────────────────────────────────────────────
Download ...
... Path './restez/downloads'
... Does path exist? 'Yes'
... N. files 486
... N. GBs 22.76
... GenBank division selections 'EST (expressed sequence tag)'
... GenBank Release 226
... Last updated '2018-09-14 10:28:13'
────────────────────────────────────────────────────────────────────────────────
Database ...
... Path './restez/sql_db'
... Does path exist? 'Yes'
... N. GBs 254.65
... Is database connected? 'Yes'
... Does the database have data? 'Yes'
... Number of sequences 76467524
... Min. sequence length 0
... Max. sequence length Inf
... Last_updated '2018-09-15 12:11:56'
> ids <- list_db_ids(n=5)
Warning message:
In list_db_ids(n = 5) : Number of ids returned was limited to [5].
Set `n=NULL` to return all ids.
> cat(gb_fasta_get(ids))
>BW190305.1 BW190305 Nori Satoh unpublished cDNA library, young adult Ciona intestinalis cDNA clone ciad090d07 5', mRNA sequence
CACATGAGCAATCTATTGAAACTTGACTTGAGCGACAACTTTATCACAAAAATAANCAAGGACGATTTTG
CAAACTGTTTTTCACTTGGTAGCATTACACTGGAAAACAACANCCTCACTCGTATAGTGCCCGGTGCATT
CANAAACACAGGGAAACTANCAAGTGTTGGATTTGTCAGAAAACCGATTAACGACGATCGAAAACGGGAC
TTTTAAAGCACTCNGTCCAAGTATAACACACATGGAGCTGTGGCA

 >BW190306.1 BW190306 Nori Satoh unpublished cDNA library, young adult Ciona intestinalis cDNA clone ciad090d09 5', mRNA sequence
AATTTGCTTTGGCAATCTTGATGCTCGTGCTTTGCGCTCACACCACCTCTGCTTGGTATGTGACATGCTC
CACCTTCGACAGCTACAAGGGTAGTATCAACAGCAGATTCAGCTCTGTTGAAAGTCGTGCCAGATCTGCA
TACAGCATCGGTCTTTGCGCCAAGAACAAGGCTTACTACGTTGCCAACGATGTCTCTTCCGTTCAATCTT
CAGTCAGCAGTTTGAGCAGCAGAGTGGGATCCGTCGAGAGCGACGTCAAGGCAGTAAAGGAGACCGTGAA
CAGCTTGAACAAATGTGCCAAGGCAGAGGATATGTTCCTTTGCATGTTCGGCATTGATTATTAAACCCTC
CCTTCATCCTTCAAGTCATGGATAGGAAACCAAGACCCTGTATCAGTGTTTTAAATACCCAGTACCATGA
CATCATATAATAAAATAAAAAAGCATCNNAAAAAAAA

 >BW190307.1 BW190307 Nori Satoh unpublished cDNA library, young adult Ciona intestinalis cDNA clone ciad090d10 5', mRNA sequence
ATCGTCTTTAAACCAAACGCGGAAATCCTTAAGCACACGATGCCTAGGGAAGACAGGAAAACATGGAAAG
CGAATTACTTCGTTAAACTTATTAACCTCTTGGATGAATATCCAAAATGCTTCATTGTCGGCGCCGACAA
CGTCAGTTCGAAGCAGATGCAACAAATTCGAACCTCTCTCCGTGAGAAGGCTGTTGTGTTGATGGGGAAA
AACACAATGATGAGGAAAGCCATCAAGGGTCATCTCAACAATAACCCGGATCTGGAGAGACTTCTCCCTC
ACATTCGTGGAAATGTTGGCTTAGTTTTCACCAAAGAAGATTTGATCGAAACCAGAGATTTACTTCTCGC
CAACAAGGTTGCTGCACCTGCCCGAGCTGGAGCCGTGGCCCCCAATGATGTCATCGTCCCCGCACAGAAC
ACTGGCTTGGGTCCCGAGAAAACTTCTTTCTTCCAAGCGTTGGCAATTCCAACCAAGATTTCAAGGGGAA
CCATTGAAATTATCAGTGACGTTCACTTGATCAAACCTGGTGATAAAGTTGGAGCTTCCGACGCAACATT
GTTGTCGATGTTGAAGATCTTCCCTTTCACTTACGGGCTTGAGATCGAACAAGTGTACGACTCTGGTAGT
GTGTTCA

 >BW190308.1 BW190308 Nori Satoh unpublished cDNA library, young adult Ciona intestinalis cDNA clone ciad090d11 5', mRNA sequence
CTTGCTTCGCACGCTTCACTAGAGACGGTGAATGAAGAAAAATTCCGTAGCAGAGCTTTTAGCGAAGGCT
CGTCTGTACTAAGCCAAGAATCAATGCAAGACCAACAACAGTCCGCCACGAGCCGTTATAAAACCGAATT
GTGTCGGCCGTTCGAGGAAAACGGAAAATGCAAATACGGCGACAAATGCCAGTTTGCTCACGGCAAACAC
GAGTTACGTAGAATGGTTCGACACCCGAAATACAAAACGGAACTTTGTCGTACGTATCACACCAGTGGTT
TTTGTCCATACGGACCACGGTGCCATTTTATCCACAACCAAGAGGACGTTGGCATTGCTAAAAAACAAAC
TCAACCTACAAGAATACAATCGCAAACGTCAGTCCCGGTAAAGATTAATAATCGCCAATTATCTCAGCAT
CAAAGACCACGTGATCTTCAACTCTTTACACAAAGCGCATCGTTCGGCCACATGCCACCTTACTCGCCAT
CCAACAGCAGCCCAAGTATTCACAACACGCAACCCATGTTTGGATTAAGTGAAGTCTTATCTCCAATTTC
ACCCATGCCACAAAATAGTACAGGCGGTCAGTTCTTTAGTTATTGTGGTAAAGGGAGTAATTCTCCGGTG
GAATCGAACCGACGTAGCAGTTTCTCCGACAACCGTCAATCCACGCCACCGATCACACAACTCGATATCT
TTAATCGAAGTCAACAACCAACTGAGTT

 >BW190309.1 BW190309 Nori Satoh unpublished cDNA library, young adult Ciona intestinalis cDNA clone ciad090d12 5', mRNA sequence
ATTCGGCACGAGGGTTTCTTGAGACCTGGATTCACAGACACTGTAGTTATGCAGTTACCGATTTGTATTG
GAACACCTACTCGTGTTGATATTATTTGTGTTCACAAACCAGTTGAAACTGAAGATCATAATTCATCTCA
CATTTCCATCATTGACATTTCACCACCTGCCAGTGTTTCCAATGCCAACAAATTCACCAAGGTGTCTTCC
TATCTTAGTAACCATTCTTTGACCAGCATAACATCTCTGGTCATCACTAAAAGTGCTCAGCTGGTAAGGC
AAGTTGTAAAATCTACAAGTAACAGCATTGAAGCTGTAAAATCGAAGCTTCAAAAGCCAGAGCCTTGCAA
ATGGGGCATCTCAACCCTTATGTTAAGAGATACTTGTTCTTGTCGAAAATGGACAGCAGTTAACATGACT
ACAAATTCAAAAACCATCACAGACGGAATTTGGGATAAAATATCTGGCAAGCGTGGTGTAGGACTTCATC
ATGGTGTCCCAACATCTTTTCCACTTGTTCAAATCCACAACAAACCACTCTTTAGTGGAACATTCCTTAA
TAAACTCACCACCGTGATTGGAATAATATTTCGTCATCATCCATTTTCATGGATATGGTCACGTGAATCT
CATCCACTTTCATCCAACTTTAGCTGCACAGAGAAGATCATTATTTCTTTCACCGTTTCTTTCCATTGTT
TACTGACCAGTGCTTGGAT
```
