### 2018-09-19 #232 @sckott
@DomBennett yeah, i think that's right for JOSS, see also the docs if you haven't already https://joss.readthedocs.io/en/latest/submitting.html
