### 2018-10-10 #232 @eveskew
Hi all, I'm back to checking out the package changes. First of all, great work Dom! It looks like you've made very thorough revisions.

Unfortunately, I'm bumping into some early housekeeping issues (most of this is for @sckott, to see how I should proceed). First, I ran into trouble with versioning of the `DBI` package. Despite the fact that the package `Imports` do specify `DBI (>= 1.0.0)`, running `devtools::install_github("AntonelliLab/restez")` did not automatically upgrade my package version (from 0.7). I was under the impression this would be the expected result, but perhaps I'm misunderstanding the behavior of `install_github()`? This caused problems in my initial code checks, but `install.packages("DBI")` resolved them. Secondly, I believe the `subprocess` package should be specified as requiring version `>= 0.8.3`. In my older version, it does not seem I had the `process_exists()` function. Finally, I'm still encountering some problems with initial code checks. For example, running `devtools::check()` ends with:

```
* preparing ‘restez’:
* creating vignettes ... ERROR
Quitting from lines 50-57 (2_search_and_fetch.Rmd)
Error: processing vignette '2_search_and_fetch.Rmd' failed with diagnostics:
Unable to execute statement 'SELECT accession,raw_definition,raw_sequence,version FROM nucleotide WHERE accession IN ('JF457164',...'.
Server says 'ParseException:SQLparser:42S02!SELECT: no such table 'nucleotide''.
Execution halted
```

I think this might just be related to a file path issue (I'm guessing Dom has an existing directory of the specified path on his machine whereas I do not), but it would take a little digging to be sure. In any case, shall I proceed, noting these minor roadblocks, or is this worth a minor update to the package before @naupaka and I keep looking?
