### 2018-10-12 #232 @sckott
`devtools::install_github` and `remotes::install_github` should install the required version of `DBI` for you. not sure what happened there.

I think that's right that `subprocess` should be `>= 0.8.3` - @DomBennett is `subprocess` only used in tests?  if so, should it be moved to Suggests?

I'm getting a failure on that vignette too.

@DomBennett could you address these few things soon so the reviewers can take another look
