### 2018-10-14 b6df7f4
more explicit ignoring of vignettes in package build
