+++ _pkgdown.yml
@@ -112,3 +112,13 @@ reference:
     - stat
     - status_class
     - testdatadir_get
+
+articles:
+- title: All articles
+  desc: Detailed tutorials for using `restez`
+  contents:
+  - '`1_rodents`'
+  - '`2_search_and_fetch`'
+  - '`3_parsing`'
+  - '`4_phylotar`'
+  - '`5_tips_and_tricks`'
