### 2018-10-14 #232 @DomBennett
Hi @sckott and @eveskew,

Hmmm.... it looks like local installs of a package do not lead to automatic updates for dependencies: https://stackoverflow.com/questions/51032658/update-dependencies-when-installing-a-local-package

`subprocess` isn't for testing, it is required for creating the spinning icon and allowing users to more easily kill a download. I've updated the DESCRIPTION to `subprocess (>= 0.8.3)` -- thanks for the spot.

As for the vignette building, I've explicitly tried to avoid building them in the R package as the vignettes depend on a database of all rodents which I build in a separate script (`other/rodent_db.R`). They should only build for me, for use in the pkgdown site, when there is a rodents database in `rodents/` in the package directory . I thought simply adding `vignettes/` to the `.Rbuildinignore` would be sufficient to prevent them being built. (For me `devtools::check()` works and vignettes are not built.)

To hopefully rectify the issue, I have more closely followed the advice from the pkgdown documentation (https://pkgdown.r-lib.org/articles/pkgdown.html#articles): dropping vignette builder in the DESCRIPTION, removing vignette details in the article yaml's and, as already done, adding vignettes/ to .Rbuildignore.

Does the `devtools::check()` still try and build the vignettes after these changes?

Thanks!
