### 2018-10-16 #232 @sckott
installs and checks fine for me know
