### 2018-11-14 9f740be
ropensci required changes
