### 2018-11-14 #232 @eveskew
Hi all!

Sorry for the late response on this, but I've looked a bit more into the revised package. @DomBennett did a very thorough job in responding to my suggestions, and I really like some of the new package functionality (the improved `restez_status()` is great!). Using the latest version of the package files, `devtools::check()` runs fine for me now.

However, I ran into a snag that I wanted to bring to Dom's attention before he migrates the package. Here's an issue I had with `db_download()` when I attempted to get some real data (the function hangs, so I ran the traceback):

```
> db_download()
───────────────────────────────────────────────────────────────
Looking up latest GenBank release ...
-
... release number 228
-
 Hide Traceback

 Rerun with Debug
 Error in FUN(X[[i]], ...) : subscript out of bounds
4.
lapply(seq_files_descripts, "[[", 2)
3.
unlist(lapply(seq_files_descripts, "[[", 2))
2.
identify_downloadable_files()
1.
db_download()
```

I believe I tracked the issue down to a couple of lines having to do with the `seq_files_descripts` object within the `identify_downloadable_files()` function. Perhaps the structure of the list has changed, because I think instead of

```
seq_files <- unlist(lapply(seq_files_descripts, '[[', 1))
descripts <- unlist(lapply(seq_files_descripts, '[[', 2))
```

you want

```
seq_files <- unlist(lapply(seq_files_descripts, '[', 1))
descripts <- unlist(lapply(seq_files_descripts, '[', 2))
```

This change at least resulted in `identify_downloadable_files()` working for me, and hopefully that's the only fix needed for `db_download()` to run. Note that I had no problems with generating and working with the demo data built into the package.

Sorry again to bring up an issue at this late hour. I know we should move along to get the package out to users, so if there are any package changes, I can be sure to review quickly and verify that `db_download()` is functional for me.

Best,
Evan
