### 2018-11-14 938c9ac
fixed out of subscript out of bounds error in download files
