### 2018-11-14 d498e95
updated version and readme
