+++ README.Rmd
@@ -19,11 +19,11 @@ assignInNamespace(x = 'custom_download2', value = restez:::custom_download,
 
 # Locally query GenBank <img src="logo.png" height="200" align="right"/>
 
-[![Build Status](https://travis-ci.org/AntonelliLab/restez.svg?branch=master)](https://travis-ci.org/AntonelliLab/restez) [![Coverage Status](https://coveralls.io/repos/github/AntonelliLab/restez/badge.svg?branch=master)](https://coveralls.io/github/AntonelliLab/restez?branch=master) [![ROpenSci status](https://badges.ropensci.org/232_status.svg)](https://github.com/ropensci/onboarding/issues/232) <!--[![CRAN downloads](http://cranlogs.r-pkg.org/badges/grand-total/restez)](https://CRAN.R-project.org/package=restez)--> [![DOI](https://zenodo.org/badge/129107980.svg)](https://zenodo.org/badge/latestdoi/129107980)
+[![Build Status](https://travis-ci.org/ropensci/restez.svg?branch=master)](https://travis-ci.org/ropensci/restez) [![Coverage Status](https://coveralls.io/repos/github/ropensci/restez/badge.svg?branch=master)](https://coveralls.io/github/ropensci/restez?branch=master) [![ROpenSci status](https://badges.ropensci.org/232_status.svg)](https://github.com/ropensci/onboarding/issues/232) <!--[![CRAN downloads](http://cranlogs.r-pkg.org/badges/grand-total/restez)](https://CRAN.R-project.org/package=restez)--> [![DOI](https://zenodo.org/badge/129107980.svg)](https://zenodo.org/badge/latestdoi/129107980)
 
 Download parts of [NCBI's GenBank](https://www.ncbi.nlm.nih.gov/nuccore) to a local folder and create a simple SQL-like database. Use 'get' tools to query the database by accession IDs. [rentrez](https://github.com/ropensci/rentrez) wrappers are available, so that if sequences are not available locally they can be searched for online through [Entrez](https://www.ncbi.nlm.nih.gov/books/NBK25500/).
 
-See the [detailed  tutorials](https://antonellilab.github.io/restez/articles/restez.html) for more information.
+See the [detailed  tutorials](https://ropensci.github.io/restez/articles/restez.html) for more information.
 
 ## Introduction
 
@@ -33,7 +33,7 @@ Downloading sequences and sequence information from GenBank and related NCBI tax
 
 This package aims to make sequence retrieval more efficient by allowing a user to download large sections of the GenBank database to their local machine and query this local database either through package specific functions or Entrez wrappers. This process is more efficient as GenBank downloads are made via NCBI's FTP using compressed sequence files. With a good internet connection and a middle-of-the-road computer, a database comprising 20 GB of sequence information can be generated in less than 10 minutes.
 
-<img src="https://raw.githubusercontent.com/AntonelliLab/restez/master/paper/outline.png" height="500" align="center"/>
+<img src="https://raw.githubusercontent.com/ropensci/restez/master/paper/outline.png" height="500" align="center"/>
 
 ## Installation
 
@@ -41,12 +41,12 @@ You can install `restez` from GitHub with:
 
 ```{r gh-installation, include=TRUE, echo=TRUE, eval=FALSE}
 # install.packages("devtools")
-devtools::install_github("AntonelliLab/restez")
+devtools::install_github("ropensci/restez")
 ```
 
 ## Quick Examples
 
-> For more detailed information on the package's functions and detailed guides on downloading, constructing and querying a database, see the [detailed  tutorials](https://antonellilab.github.io/restez/articles/restez.html).
[…]
