+++ DESCRIPTION
@@ -1,14 +1,14 @@
 Package: restez
 Type: Package
 Title: Create and Query a Local Copy of GenBank in R
-Version: 0.1.0
+Version: 1.0.0
 Authors@R: person("Dom", "Bennett", role = c("aut", "cre"), email = "dominic.john.bennett@gmail.com")
 Maintainer: D.J. Bennett <dominic.john.bennett@gmail.com>
 Description: Download large sections of GenBank and generate a local SQL-based
     database. A user can then query this database using restez functions or
     through rentrez wrappers.
-URL: https://github.com/AntonelliLab/restez#readme
-BugReports: https://github.com/AntonelliLab/restez/issues
+URL: https://github.com/ropensci/restez#readme
+BugReports: https://github.com/ropensci/restez/issues
 License: MIT + file LICENSE
 Encoding: UTF-8
 LazyData: true
