### 2018-11-14 #232 @DomBennett
@sckott I added the restez repo to the [restez team](https://github.com/orgs/ropensci/teams/restez). I think I need to be a member of that to push, right? At least, I don't seem to be able to add myself to that team.
