+++ R/print-tools.R
@@ -30,20 +30,20 @@ cat_line <- function(...) {
   cat(paste0(..., "\n"), sep = "")
 }
 
-#' @name line_spinner
-#' @title Spin the lines
-#' @description Prints cool little spinning lines to console to indicate a
-#' running process. Used in conjunction with a while loop.
-#' @return NULL
-#' @family private
-line_spinner <- function() {
-  # switch to lines because dots stopped working in RStudio
-  #sp <- cli::get_spinner(which = 'line')
-  interval <- 130/1000
-  frames <- c("\\", "|", "/", "-")
-  for (i in seq_along(frames) - 1) {
-    fr <- unclass(frames[i %% length(frames) + 1])
-    cat("\r", crayon::red(fr), sep = "")
-    Sys.sleep(interval)
-  }
-}
+# #' @name line_spinner
+# #' @title Spin the lines
+# #' @description Prints cool little spinning lines to console to indicate a
+# #' running process. Used in conjunction with a while loop.
+# #' @return NULL
+# #' @family private
+# line_spinner <- function() {
+#   # switch to lines because dots stopped working in RStudio
+#   #sp <- cli::get_spinner(which = 'line')
+#   interval <- 130/1000
+#   frames <- c("\\", "|", "/", "-")
+#   for (i in seq_along(frames) - 1) {
+#     fr <- unclass(frames[i %% length(frames) + 1])
+#     cat("\r", crayon::red(fr), sep = "")
+#     Sys.sleep(interval)
+#   }
+# }
