+++ man/db_sqlngths_get.Rd
@@ -24,7 +24,6 @@ Other private: \code{\link{add_rcrd_log}},
   \code{\link{custom_download2}},
   \code{\link{custom_download}},
   \code{\link{db_sqlngths_log}}, \code{\link{dir_size}},
-  \code{\link{download_cmd_generate}},
   \code{\link{dwnld_path_get}},
   \code{\link{dwnld_rcrd_log}},
   \code{\link{entrez_fasta_get}},
@@ -47,20 +46,18 @@ Other private: \code{\link{add_rcrd_log}},
   \code{\link{gb_sql_query}},
   \code{\link{gbrelease_check}},
   \code{\link{gbrelease_get}}, \code{\link{gbrelease_log}},
-  \code{\link{handle_run}}, \code{\link{has_data}},
+  \code{\link{has_data}},
   \code{\link{identify_downloadable_files}},
   \code{\link{last_add_get}}, \code{\link{last_dwnld_get}},
   \code{\link{last_entry_get}},
   \code{\link{latest_genbank_release_notes}},
   \code{\link{latest_genbank_release}},
-  \code{\link{line_spinner}},
   \code{\link{message_missing}}, \code{\link{mock_def}},
   \code{\link{mock_gb_df_generate}},
   \code{\link{mock_org}}, \code{\link{mock_rec}},
   \code{\link{mock_seq}}, \code{\link{predict_datasizes}},
   \code{\link{print.status}}, \code{\link{readme_log}},
   \code{\link{restez_path_check}}, \code{\link{restez_rl}},
-  \code{\link{rhandle_generate}},
   \code{\link{seshinfo_log}}, \code{\link{setup}},
   \code{\link{slctn_get}}, \code{\link{slctn_log}},
   \code{\link{sql_path_get}}, \code{\link{status_class}},

[…]
