### 2018-11-14 #232 @eveskew
@DomBennett, just wanted to report that I cloned the rOpenSci version of the package, and all seems to be well now! I was able to download a sample of real data (phage sequences), create the database, and query it successfully using the suite of `gb_*_get()` functions. Great work!
