### 2018-11-15 e1c6428
callr with gb_build, tests passing
