+++ examples/restez_ready.R
@@ -1,8 +1,8 @@
 library(restez)
 fp <- tempdir()
 restez_path_set(filepath = fp)
-restez_connect()
 demo_db_create()
+restez_connect()
 (restez_ready())
 restez_disconnect()
 db_delete(everything = TRUE)
