+++ examples/gb_version_get.R
@@ -1,7 +1,7 @@
 library(restez)
 restez_path_set(filepath = tempdir())
-restez_connect()
 demo_db_create()
+restez_connect()
 (ver <- gb_version_get(id = 'demo_1'))
 (vers <- gb_version_get(id = c('demo_1', 'demo_2')))
 
