+++ tests/testthat/test-db-tools.R
@@ -18,6 +18,7 @@ test_that('demo_db_create() works', {
   restez:::setup()
   on.exit(restez:::cleanup())
   demo_db_create()
+  restez_connect()
   sequence <- gb_sequence_get('demo_1')[[1]]
   expect_true(grepl(pattern = '[atcg]', x = sequence, ignore.case = TRUE))
 })
