+++ tests/testthat/test-rentrez-wrappers.R
@@ -13,6 +13,7 @@ records <- readRDS(file = file.path(data_d, 'records.RData'))
 # SETUP
 restez:::cleanup()
 restez:::setup()
+restez_connect()
 df <- restez:::gb_df_generate(records = sample(records, size = nrcrds))
 ids <- as.character(df[['accession']])
 restez:::gb_sql_add(df = df)
