+++ NEWS.md
@@ -1,0 +1,5 @@
+# outsider 1.0.0
+
+## Post-review version of `restez` released.
+
+* Download and query parts of GenBank from within R
