+++ other/rodent_db.R
@@ -8,4 +8,3 @@ restez_path_set(rodents_path)
 # download for domain 15
 dwntm <- system.time(db_download(preselection = '15'))
 crttm <- system.time(db_create(min_length = 100, max_length = 1000))
-# always disconnect
