+++ _pkgdown.yml
@@ -53,7 +53,6 @@ reference:
     - db_sqlngths_get
     - db_sqlngths_log
     - dir_size
-    - download_cmd_generate
     - dwnld_path_get
     - dwnld_rcrd_log
     - entrez_fasta_get
@@ -74,6 +73,8 @@ reference:
     - file_download
     - filename_log
     - flatfile_read
+    - gb_build
+    - gb_build2
     - gb_df_create
     - gb_df_generate
     - gb_sql_add
@@ -81,7 +82,6 @@ reference:
     - gbrelease_check
     - gbrelease_get
     - gbrelease_log
-    - handle_run
     - has_data
     - identify_downloadable_files
     - last_add_get
@@ -91,7 +91,6 @@ reference:
     - latest_genbank_release
     - latest_genbank_release_notes
     - latest_genbank_release_notes
-    - line_spinner
     - message_missing
     - mock_def
     - mock_gb_df_generate
@@ -100,10 +99,10 @@ reference:
     - mock_seq
     - predict_datasizes
     - print.status
+    - quiet_connect
     - readme_log
     - restez_path_check
     - restez_rl
-    - rhandle_generate
     - seshinfo_log
     - setup
     - slctn_get
