+++ README.md
@@ -55,7 +55,6 @@ restez_path_set(rstz_pth)
 db_download()
 # after download, create the local database
 db_create()
-#> Remember to run `restez_disconnect()`
 ```
 
 ### Query
@@ -72,68 +71,54 @@ id <- sample(list_db_ids(), 1)
 # sequences
 seq <- gb_sequence_get(id)[[1]]
 str(seq)
-#>  chr "ATCAGAGACTTAGGACAGAACAGTCAGCGACAAACGCAAGGAAATTCTTTCTCTCCTTCCTTTCTTTCTGATTGTTTCTTCGTTCGCGGTAAAACTCACAAGTTTGCGTAA"| __truncated__
+#>  chr "GATCCATATGTATACGGAAAGCGTGTTTACCTCTTTGGTCAGATANTAACTGATGCTCACCCAATCNTNCCGTAGACCGTGCCTGATTGGGCCGGANGACCTGGGGAACGA"| __truncated__
 # definitions
 def <- gb_definition_get(id)[[1]]
 print(def)
-#> [1] "Unidentified Cotton leaf curl Rajasthan virus-associated DNA clone pNDM1.5, partial sequence"
+#> [1] "Unidentified clone A3 DNA sequence from ocean beach sand"
 # organisms
 org <- gb_organism_get(id)[[1]]
 print(org)
-#> [1] "unidentified Cotton leaf curl Rajasthan virus-associated DNA"
+#> [1] "unidentified"
 # or whole records
 rec <- gb_record_get(id)[[1]]
 cat(rec)
-#> LOCUS       DQ415960                1129 bp    DNA     linear   UNA 07-MAY-2006
-#> DEFINITION  Unidentified Cotton leaf curl Rajasthan virus-associated DNA clone
-#>             pNDM1.5, partial sequence.
-#> ACCESSION   DQ415960
-#> VERSION     DQ415960.1
+#> LOCUS       AF298085                 537 bp    DNA     linear   UNA 23-NOV-2000
+#> DEFINITION  Unidentified clone A3 DNA sequence from ocean beach sand.
+#> ACCESSION   AF298085
+#> VERSION     AF298085.1
 #> KEYWORDS    .
-#> SOURCE      unidentified Cotton leaf curl Rajasthan virus-associated DNA
-#>   ORGANISM  unidentified Cotton leaf curl Rajasthan virus-associated DNA
+#> SOURCE      unidentified
+#>   ORGANISM  unidentified
 #>             unclassified sequences.
-#> REFERENCE   1  (bases 1 to 1129)
-#>   AUTHORS   Radhakrishnan,G., Malathi,V.G. and Varma,A.
-#>   TITLE     Cotton leaf curl Rajasthan virus associated novel DNA molecules
-#>             (NDMs)
+#> REFERENCE   1  (bases 1 to 537)
+#>   AUTHORS   Naviaux,R.K.
+#>   TITLE     Sand DNA: a multigenomic library on the beach
 #>   JOURNAL   Unpublished
-#> REFERENCE   2  (bases 1 to 1129)
-#>   AUTHORS   Radhakrishnan,G., Malathi,V.G. and Varma,A.
[…]
