### 2018-11-15 1133812
attributes added, new tests, check passing
