+++ DESCRIPTION
@@ -2,7 +2,9 @@ Package: restez
 Type: Package
 Title: Create and Query a Local Copy of GenBank in R
 Version: 1.0.0
-Authors@R: person("Dom", "Bennett", role = c("aut", "cre"), email = "dominic.john.bennett@gmail.com")
+Authors@R: person("Dom", "Bennett", role = c("aut", "cre"),
+    email = "dominic.john.bennett@gmail.com",
+    comment = c(ORCID = "0000-0003-2722-1359"))
 Maintainer: D.J. Bennett <dominic.john.bennett@gmail.com>
 Description: Download large sections of GenBank and generate a local SQL-based
     database. A user can then query this database using restez functions or
