### 2018-11-15 #232 @naupaka
@DomBennett I just re-tried to set up the plant/fungal DB, and it hung on the `db_create()` step. After ^C, it returned the following error:

```
... 'gbpln231.seq.gz'(148 / 232)
... 'gbpln232.seq.gz'(149 / 232)
... 'gbpln24.seq.gz'(150 / 232)
... 'gbpln25.seq.gz'(151 / 232)
^CError in .local(conn, statement, ...) :
  Unable to execute statement 'select schemas.name as sn, tables.name as tn from sys.tables join sys.schemas on tables.schema_id=sc...'.
Server says 'MALException:dataflow:DFLOWscheduler(): q_dequeue(flow->done) returned NULL'.
```

Is this the same error that you were talking about earlier?
