### 2018-11-15 f00ecca
changes for CRAN
