+++ DESCRIPTION
@@ -5,7 +5,7 @@ Version: 1.0.0
 Authors@R: person("Dom", "Bennett", role = c("aut", "cre"),
     email = "dominic.john.bennett@gmail.com",
     comment = c(ORCID = "0000-0003-2722-1359"))
-Maintainer: D.J. Bennett <dominic.john.bennett@gmail.com>
+Maintainer: Dom Bennett <dominic.john.bennett@gmail.com>
 Description: Download large sections of GenBank and generate a local SQL-based
     database. A user can then query this database using restez functions or
     through rentrez wrappers.
@@ -19,7 +19,7 @@ Depends:
 Imports:
     utils,
     rentrez,
-    MonetDBLite,
+    MonetDBLite (>= 0.6.0),
     DBI (>= 1.0.0),
     downloader,
     RCurl,
