### 2018-11-15 c978598
sped up examples for CRAN
