+++ examples/count_db_ids.R
@@ -1,6 +1,6 @@
 library(restez)
 restez_path_set(filepath = tempdir())
-demo_db_create()
+demo_db_create(n = 5)
 restez_connect()
 (count_db_ids())
 
