+++ examples/restez_ready.R
@@ -1,7 +1,7 @@
 library(restez)
 fp <- tempdir()
 restez_path_set(filepath = fp)
-demo_db_create()
+demo_db_create(n = 5)
 restez_connect()
 (restez_ready())
 restez_disconnect()
