+++ examples/gb_fasta_get.R
@@ -1,6 +1,6 @@
 library(restez)
 restez_path_set(filepath = tempdir())
-demo_db_create()
+demo_db_create(n = 5)
 restez_connect()
 (fasta <- gb_fasta_get(id = 'demo_1'))
 (fastas <- gb_fasta_get(id = c('demo_1', 'demo_2')))
