+++ examples/is_in_db.R
@@ -2,7 +2,7 @@ library(restez)
 # set the restez path to a temporary dir
 restez_path_set(filepath = tempdir())
 # create demo database
-demo_db_create(n = 100)
+demo_db_create(n = 5)
 restez_connect()
 # in the demo, IDs are 'demo_1', 'demo_2' ...
 ids <- c('thisisnotanid', 'demo_1', 'demo_2')
