+++ examples/gb_version_get.R
@@ -1,6 +1,6 @@
 library(restez)
 restez_path_set(filepath = tempdir())
-demo_db_create()
+demo_db_create(n = 5)
 restez_connect()
 (ver <- gb_version_get(id = 'demo_1'))
 (vers <- gb_version_get(id = c('demo_1', 'demo_2')))
