+++ examples/gb_organism_get.R
@@ -1,6 +1,6 @@
 library(restez)
 restez_path_set(filepath = tempdir())
-demo_db_create()
+demo_db_create(n = 5)
 restez_connect()
 (org <- gb_organism_get(id = 'demo_1'))
 (orgs <- gb_organism_get(id = c('demo_1', 'demo_2')))
