+++ examples/gb_sequence_get.R
@@ -1,6 +1,6 @@
 library(restez)
 restez_path_set(filepath = tempdir())
-demo_db_create()
+demo_db_create(n = 5)
 restez_connect()
 (seq <- gb_sequence_get(id = 'demo_1'))
 (seqs <- gb_sequence_get(id = c('demo_1', 'demo_2')))
