### 2018-11-15 #232 @DomBennett
@naupaka Hmmm... looks like a new error I will investigate it!

I have in the past had issues with the database but haven't always been able to recreate the error. You might find that re-running the `db_create` function will have no problems a second time. Unfortunately, it would have to start from the beginning again.

Also note, I have just updated the `db_create` function. It no longer requires `restez_connect` before running. This is so I can run most of the bulk of the function with `callr` -- makes killing the running process easier.

(The database behaviour is not 100% predictable, at least for me. For example, I found I couldn't build a database on a USB stick, I should add that to the documentation.)
