### 2018-11-15 9558265
minor cran changes
