+++ DESCRIPTION
@@ -1,14 +1,16 @@
 Package: restez
 Type: Package
-Title: Create and Query a Local Copy of GenBank in R
+Title: Create and Query a Local Copy of 'GenBank' in R
 Version: 1.0.0
 Authors@R: person("Dom", "Bennett", role = c("aut", "cre"),
     email = "dominic.john.bennett@gmail.com",
     comment = c(ORCID = "0000-0003-2722-1359"))
 Maintainer: Dom Bennett <dominic.john.bennett@gmail.com>
-Description: Download large sections of GenBank and generate a local SQL-based
-    database. A user can then query this database using restez functions or
-    through rentrez wrappers.
+Description: Download large sections of
+    'GenBank' <https://www.ncbi.nlm.nih.gov/genbank/> and generate a local
+    SQL-based database. A user can then query this database using 'restez'
+    functions or through 'rentrez' <https://CRAN.R-project.org/package=rentrez>
+    wrappers.
 URL: https://github.com/ropensci/restez#readme
 BugReports: https://github.com/ropensci/restez/issues
 License: MIT + file LICENSE
