### 2018-11-15 ff39219
minor print changes
