### 2018-11-16 #232 @DomBennett
Hi @naupaka,

The plant/fungal database seems to work for me -- see attached.

Let me know if you're still having problems.

[plants_fungi_resetz_rsession.txt](https://github.com/ropensci/onboarding/files/2588605/plants_fungi_resetz_rsession.txt)
[plants_fungi_restez_log.txt](https://github.com/ropensci/onboarding/files/2588606/plants_fungi_restez_log.txt)
