+++ DESCRIPTION
@@ -34,4 +34,4 @@ Suggests:
     knitr,
     R.utils,
     rmarkdown
-RoxygenNote: 6.1.0
+RoxygenNote: 6.1.1
