### 2018-11-19 #232 @naupaka
Awesome! All works great for me now. Thanks for all your hard work on this, @DomBennett! Going to be a really helpful tool.
