### 2018-11-19 #232 @sckott
not a dumb question.  You do have to submit it manually (maybe someday it will be automated).

* For JOSS:
    * Looks like there's already a Zenodo web hook, so that's good.
    * If you haven't already, generate a new release associated with a git tag in your repo. This doesn't mean you have to submit to CRAN with that git tag/release, its just to get a Zenodo DOI - and you can put a zenodo badge in your readme, then after each git tag it will generate a new Zenodo release
    * Then submit your `paper.md` to JOSS at http://joss.theoj.org/papers/new
