### 2018-11-21 87501dd
Update NEWS.md
