+++ NEWS.md
@@ -1,4 +1,4 @@
-# outsider 1.0.0
+# restez 1.0.0
 
 ## Post-review version of `restez` released.
 
