### 2018-11-26 #232 @DomBennett
Thanks for the helpful answer! I've now submitted the paper -- I was waiting for CRAN to host.
