### 2018-11-26 #232 @sckott
cool, closing now, seems like JOSS and blog post are started
